Network.AWS.SSM.CreateActivation

Description

Registers your on-premises server or virtual machine with Amazon EC2 so that you can manage these resources using Run Command. An on-premises server or virtual machine that has been registered with EC2 is called a managed instance. For more information about activations, see Setting Up Systems Manager in Hybrid Environments .

Creating a Request

createActivation

Arguments

:: Text
caIAMRole

-> CreateActivation

Creates a value of CreateActivation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

caDefaultInstanceName - The name of the registered, managed instance as it will appear in the Amazon EC2 console or when you use the AWS command line tools to list EC2 resources.

caRegistrationLimit - Specify the maximum number of managed instances you want to register. The default value is 1 instance.

caExpirationDate - The date by which this activation request should expire. The default value is 24 hours.

caDescription - A userdefined description of the resource that you want to register with Amazon EC2.

caIAMRole - The Amazon Identity and Access Management (IAM) role that you want to assign to the managed instance.

data CreateActivation

See: createActivation smart constructor.

Request Lenses

caDefaultInstanceName :: Lens' CreateActivation (Maybe Text)

The name of the registered, managed instance as it will appear in the Amazon EC2 console or when you use the AWS command line tools to list EC2 resources.

caRegistrationLimit :: Lens' CreateActivation (Maybe Natural)

Specify the maximum number of managed instances you want to register. The default value is 1 instance.

caExpirationDate :: Lens' CreateActivation (Maybe UTCTime)

The date by which this activation request should expire. The default value is 24 hours.

caDescription :: Lens' CreateActivation (Maybe Text)

A userdefined description of the resource that you want to register with Amazon EC2.

caIAMRole :: Lens' CreateActivation Text

The Amazon Identity and Access Management (IAM) role that you want to assign to the managed instance.

Destructuring the Response

createActivationResponse

Arguments

:: Int
carsResponseStatus

-> CreateActivationResponse

Creates a value of CreateActivationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

carsActivationId - The ID number generated by the system when it processed the activation. The activation ID functions like a user name.

carsActivationCode - The code the system generates when it processes the activation. The activation code functions like a password to validate the activation ID.

carsResponseStatus - -- | The response status code.

data CreateActivationResponse

See: createActivationResponse smart constructor.

Response Lenses

carsActivationId :: Lens' CreateActivationResponse (Maybe Text)

The ID number generated by the system when it processed the activation. The activation ID functions like a user name.

carsActivationCode :: Lens' CreateActivationResponse (Maybe Text)

The code the system generates when it processes the activation. The activation code functions like a password to validate the activation ID.

carsResponseStatus :: Lens' CreateActivationResponse Int

- | The response status code.