Network.AWS.SSM.CreateAssociation

Description

Associates the specified Systems Manager document with the specified instances or targets.

When you associate a document with one or more instances using instance IDs or tags, the SSM Agent running on the instance processes the document and configures the instance as specified.

If you associate a document with an instance that already has an associated document, the system throws the AssociationAlreadyExists exception.

Creating a Request

createAssociation

Arguments

:: Text
caName

-> CreateAssociation

Creates a value of CreateAssociation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

caInstanceId - The instance ID.

caScheduleExpression - A cron expression when the association will be applied to the target(s).

caOutputLocation - An Amazon S3 bucket where you want to store the output details of the request.

caTargets - The targets (either instances or tags) for the association.

caParameters - The parameters for the documents runtime configuration.

caDocumentVersion - The document version you want to associate with the target(s). Can be a specific version or the default version.

caName - The name of the Systems Manager document.

data CreateAssociation

See: createAssociation smart constructor.

Request Lenses

caInstanceId :: Lens' CreateAssociation (Maybe Text)

The instance ID.

caScheduleExpression :: Lens' CreateAssociation (Maybe Text)

A cron expression when the association will be applied to the target(s).

caOutputLocation :: Lens' CreateAssociation (Maybe InstanceAssociationOutputLocation)

An Amazon S3 bucket where you want to store the output details of the request.

caTargets :: Lens' CreateAssociation [Target]

The targets (either instances or tags) for the association.

caParameters :: Lens' CreateAssociation (HashMap Text [Text])

The parameters for the documents runtime configuration.

caDocumentVersion :: Lens' CreateAssociation (Maybe Text)

The document version you want to associate with the target(s). Can be a specific version or the default version.

caName :: Lens' CreateAssociation Text

The name of the Systems Manager document.

Destructuring the Response

createAssociationResponse

Arguments

:: Int
crsResponseStatus

-> CreateAssociationResponse

Creates a value of CreateAssociationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crsAssociationDescription - Information about the association.

crsResponseStatus - -- | The response status code.

data CreateAssociationResponse

See: createAssociationResponse smart constructor.

Response Lenses

crsAssociationDescription :: Lens' CreateAssociationResponse (Maybe AssociationDescription)

Information about the association.

crsResponseStatus :: Lens' CreateAssociationResponse Int

- | The response status code.