Network.AWS.SSM.CreateAssociationBatch

Description

Associates the specified configuration documents with the specified instances.

When you associate a configuration document with an instance, the configuration agent on the instance processes the configuration document and configures the instance as specified.

If you associate a configuration document with an instance that already has an associated configuration document, we replace the current configuration document with the new configuration document.

http://docs.aws.amazon.com/ssm/latest/APIReference/API_CreateAssociationBatch.html

Request

data CreateAssociationBatch

Request constructor

createAssociationBatch :: CreateAssociationBatch

CreateAssociationBatch constructor.

The fields accessible through corresponding lenses are:

cabEntries :: [CreateAssociationBatchRequestEntry]

Request lenses

cabEntries :: Lens' CreateAssociationBatch [CreateAssociationBatchRequestEntry]

One or more associations.

Response

data CreateAssociationBatchResponse

Response constructor

createAssociationBatchResponse :: CreateAssociationBatchResponse

CreateAssociationBatchResponse constructor.

The fields accessible through corresponding lenses are:

cabrFailed :: [FailedCreateAssociation]

cabrSuccessful :: [AssociationDescription]

Response lenses

cabrFailed :: Lens' CreateAssociationBatchResponse [FailedCreateAssociation]

Information about the associations that failed.

cabrSuccessful :: Lens' CreateAssociationBatchResponse [AssociationDescription]

Information about the associations that succeeded.