Network.AWS.SSM.CreateDocument

Description

Creates a configuration document.

After you create a configuration document, you can use CreateAssociation to associate it with one or more running instances.

http://docs.aws.amazon.com/ssm/latest/APIReference/API_CreateDocument.html

Request

data CreateDocument

Request constructor

createDocument

Arguments

:: Text
cdContent

-> Text
cdName

-> CreateDocument

CreateDocument constructor.

The fields accessible through corresponding lenses are:

cdContent :: Text

cdName :: Text

Request lenses

cdContent :: Lens' CreateDocument Text

A valid JSON file. For more information about the contents of this file, see Configuration Document.

cdName :: Lens' CreateDocument Text

A name for the configuration document.

Response

data CreateDocumentResponse

Response constructor

createDocumentResponse :: CreateDocumentResponse

CreateDocumentResponse constructor.

The fields accessible through corresponding lenses are:

cdrDocumentDescription :: Maybe DocumentDescription

Response lenses

cdrDocumentDescription :: Lens' CreateDocumentResponse (Maybe DocumentDescription)

Information about the configuration document.