Network.AWS.SSM.CreateMaintenanceWindow

Description

Creates a new Maintenance Window.

Creating a Request

createMaintenanceWindow

Arguments

:: Text
cmwName

-> Text
cmwSchedule

-> Natural
cmwDuration

-> Natural
cmwCutoff

-> Bool
cmwAllowUnassociatedTargets

-> CreateMaintenanceWindow

Creates a value of CreateMaintenanceWindow with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cmwClientToken - User-provided idempotency token.

cmwDescription - An optional description for the Maintenance Window. We recommend specifying a description to help your organize your Maintenance Windows.

cmwName - The name of the Maintenance Window.

cmwSchedule - The schedule of the Maintenance Window in the form of a cron or rate expression.

cmwDuration - The duration of the Maintenance Window in hours.

cmwCutoff - The number of hours before the end of the Maintenance Window that Systems Manager stops scheduling new tasks for execution.

cmwAllowUnassociatedTargets - Whether targets must be registered with the Maintenance Window before tasks can be defined for those targets.

data CreateMaintenanceWindow

See: createMaintenanceWindow smart constructor.

Request Lenses

cmwClientToken :: Lens' CreateMaintenanceWindow (Maybe Text)

User-provided idempotency token.

cmwDescription :: Lens' CreateMaintenanceWindow (Maybe Text)

An optional description for the Maintenance Window. We recommend specifying a description to help your organize your Maintenance Windows.

cmwName :: Lens' CreateMaintenanceWindow Text

The name of the Maintenance Window.

cmwSchedule :: Lens' CreateMaintenanceWindow Text

The schedule of the Maintenance Window in the form of a cron or rate expression.

cmwDuration :: Lens' CreateMaintenanceWindow Natural

The duration of the Maintenance Window in hours.

cmwCutoff :: Lens' CreateMaintenanceWindow Natural

The number of hours before the end of the Maintenance Window that Systems Manager stops scheduling new tasks for execution.

cmwAllowUnassociatedTargets :: Lens' CreateMaintenanceWindow Bool

Whether targets must be registered with the Maintenance Window before tasks can be defined for those targets.

Destructuring the Response

createMaintenanceWindowResponse

Arguments

:: Int
cmwrsResponseStatus

-> CreateMaintenanceWindowResponse

Creates a value of CreateMaintenanceWindowResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cmwrsWindowId - The ID of the created Maintenance Window.

cmwrsResponseStatus - -- | The response status code.

data CreateMaintenanceWindowResponse

See: createMaintenanceWindowResponse smart constructor.

Response Lenses

cmwrsWindowId :: Lens' CreateMaintenanceWindowResponse (Maybe Text)

The ID of the created Maintenance Window.

cmwrsResponseStatus :: Lens' CreateMaintenanceWindowResponse Int

- | The response status code.