Network.AWS.SSM.CreatePatchBaseline

Description

Creates a patch baseline.

Creating a Request

createPatchBaseline

Arguments

:: Text
cpbName

-> CreatePatchBaseline

Creates a value of CreatePatchBaseline with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpbApprovalRules - A set of rules used to include patches in the baseline.

cpbClientToken - User-provided idempotency token.

cpbOperatingSystem - Defines the operating system the patch baseline applies to. Supported operating systems include WINDOWS, AMAZON_LINUX, UBUNTU and REDHAT_ENTERPRISE_LINUX. The Default value is WINDOWS.

cpbGlobalFilters - A set of global filters used to exclude patches from the baseline.

cpbApprovedPatchesComplianceLevel - Defines the compliance level for approved patches. This means that if an approved patch is reported as missing, this is the severity of the compliance violation. Valid compliance severity levels include the following: CRITICAL, HIGH, MEDIUM, LOW, INFORMATIONAL, UNSPECIFIED. The default value is UNSPECIFIED.

cpbApprovedPatches - A list of explicitly approved patches for the baseline.

cpbRejectedPatches - A list of explicitly rejected patches for the baseline.

cpbDescription - A description of the patch baseline.

cpbName - The name of the patch baseline.

data CreatePatchBaseline

See: createPatchBaseline smart constructor.

Request Lenses

cpbApprovalRules :: Lens' CreatePatchBaseline (Maybe PatchRuleGroup)

A set of rules used to include patches in the baseline.

cpbClientToken :: Lens' CreatePatchBaseline (Maybe Text)

User-provided idempotency token.

cpbOperatingSystem :: Lens' CreatePatchBaseline (Maybe OperatingSystem)

Defines the operating system the patch baseline applies to. Supported operating systems include WINDOWS, AMAZON_LINUX, UBUNTU and REDHAT_ENTERPRISE_LINUX. The Default value is WINDOWS.

cpbGlobalFilters :: Lens' CreatePatchBaseline (Maybe PatchFilterGroup)

A set of global filters used to exclude patches from the baseline.

cpbApprovedPatchesComplianceLevel :: Lens' CreatePatchBaseline (Maybe PatchComplianceLevel)

Defines the compliance level for approved patches. This means that if an approved patch is reported as missing, this is the severity of the compliance violation. Valid compliance severity levels include the following: CRITICAL, HIGH, MEDIUM, LOW, INFORMATIONAL, UNSPECIFIED. The default value is UNSPECIFIED.

cpbApprovedPatches :: Lens' CreatePatchBaseline [Text]

A list of explicitly approved patches for the baseline.

cpbRejectedPatches :: Lens' CreatePatchBaseline [Text]

A list of explicitly rejected patches for the baseline.

cpbDescription :: Lens' CreatePatchBaseline (Maybe Text)

A description of the patch baseline.

cpbName :: Lens' CreatePatchBaseline Text

The name of the patch baseline.

Destructuring the Response

createPatchBaselineResponse

Arguments

:: Int
cpbrsResponseStatus

-> CreatePatchBaselineResponse

Creates a value of CreatePatchBaselineResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpbrsBaselineId - The ID of the created patch baseline.

cpbrsResponseStatus - -- | The response status code.

data CreatePatchBaselineResponse

See: createPatchBaselineResponse smart constructor.

Response Lenses

cpbrsBaselineId :: Lens' CreatePatchBaselineResponse (Maybe Text)

The ID of the created patch baseline.

cpbrsResponseStatus :: Lens' CreatePatchBaselineResponse Int

- | The response status code.