Network.AWS.SSM.CreateResourceDataSync

Description

Creates a resource data sync configuration to a single bucket in Amazon S3. This is an asynchronous operation that returns immediately. After a successful initial sync is completed, the system continuously syncs data to the Amazon S3 bucket. To check the status of the sync, use the ListResourceDataSync operation.

By default, data is not encrypted in Amazon S3. We strongly recommend that you enable encryption in Amazon S3 to ensure secure data storage. We also recommend that you secure access to the Amazon S3 bucket by creating a restrictive bucket policy. To view an example of a restrictive Amazon S3 bucket policy for Resource Data Sync, see Configuring Resource Data Sync for Inventory .

Creating a Request

createResourceDataSync

Arguments

:: Text
crdsSyncName

-> ResourceDataSyncS3Destination
crdsS3Destination

-> CreateResourceDataSync

Creates a value of CreateResourceDataSync with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crdsSyncName - A name for the configuration.

crdsS3Destination - Amazon S3 configuration details for the sync.

data CreateResourceDataSync

See: createResourceDataSync smart constructor.

Request Lenses

crdsSyncName :: Lens' CreateResourceDataSync Text

A name for the configuration.

crdsS3Destination :: Lens' CreateResourceDataSync ResourceDataSyncS3Destination

Amazon S3 configuration details for the sync.

Destructuring the Response

createResourceDataSyncResponse

Arguments

:: Int
crdsrsResponseStatus

-> CreateResourceDataSyncResponse

Creates a value of CreateResourceDataSyncResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crdsrsResponseStatus - -- | The response status code.

data CreateResourceDataSyncResponse

See: createResourceDataSyncResponse smart constructor.

Response Lenses

crdsrsResponseStatus :: Lens' CreateResourceDataSyncResponse Int

- | The response status code.