Network.AWS.SSM.DeleteDocument

Description

Deletes the specified configuration document.

You must use DeleteAssociation to disassociate all instances that are associated with the configuration document before you can delete it.

http://docs.aws.amazon.com/ssm/latest/APIReference/API_DeleteDocument.html

Request

data DeleteDocument

Request constructor

deleteDocument

Arguments

:: Text
dd2Name

-> DeleteDocument

DeleteDocument constructor.

The fields accessible through corresponding lenses are:

dd2Name :: Text

Request lenses

dd2Name :: Lens' DeleteDocument Text

The name of the configuration document.

Response

data DeleteDocumentResponse

Response constructor

deleteDocumentResponse :: DeleteDocumentResponse

DeleteDocumentResponse constructor.