Network.AWS.SSM.DeleteMaintenanceWindow

Description

Deletes a Maintenance Window.

Creating a Request

deleteMaintenanceWindow

Arguments

:: Text
dmwWindowId

-> DeleteMaintenanceWindow

Creates a value of DeleteMaintenanceWindow with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmwWindowId - The ID of the Maintenance Window to delete.

data DeleteMaintenanceWindow

See: deleteMaintenanceWindow smart constructor.

Request Lenses

dmwWindowId :: Lens' DeleteMaintenanceWindow Text

The ID of the Maintenance Window to delete.

Destructuring the Response

deleteMaintenanceWindowResponse

Arguments

:: Int
dmwrsResponseStatus

-> DeleteMaintenanceWindowResponse

Creates a value of DeleteMaintenanceWindowResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmwrsWindowId - The ID of the deleted Maintenance Window.

dmwrsResponseStatus - -- | The response status code.

data DeleteMaintenanceWindowResponse

See: deleteMaintenanceWindowResponse smart constructor.

Response Lenses

dmwrsWindowId :: Lens' DeleteMaintenanceWindowResponse (Maybe Text)

The ID of the deleted Maintenance Window.

dmwrsResponseStatus :: Lens' DeleteMaintenanceWindowResponse Int

- | The response status code.