Network.AWS.SSM.DeregisterPatchBaselineForPatchGroup

Description

Removes a patch group from a patch baseline.

Creating a Request

deregisterPatchBaselineForPatchGroup

Arguments

:: Text
dpbfpgBaselineId

-> Text
dpbfpgPatchGroup

-> DeregisterPatchBaselineForPatchGroup

Creates a value of DeregisterPatchBaselineForPatchGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dpbfpgBaselineId - The ID of the patch baseline to deregister the patch group from.

dpbfpgPatchGroup - The name of the patch group that should be deregistered from the patch baseline.

data DeregisterPatchBaselineForPatchGroup

See: deregisterPatchBaselineForPatchGroup smart constructor.

Request Lenses

dpbfpgBaselineId :: Lens' DeregisterPatchBaselineForPatchGroup Text

The ID of the patch baseline to deregister the patch group from.

dpbfpgPatchGroup :: Lens' DeregisterPatchBaselineForPatchGroup Text

The name of the patch group that should be deregistered from the patch baseline.

Destructuring the Response

deregisterPatchBaselineForPatchGroupResponse

Arguments

:: Int
dpbfpgrsResponseStatus

-> DeregisterPatchBaselineForPatchGroupResponse

Creates a value of DeregisterPatchBaselineForPatchGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dpbfpgrsBaselineId - The ID of the patch baseline the patch group was deregistered from.

dpbfpgrsPatchGroup - The name of the patch group deregistered from the patch baseline.

dpbfpgrsResponseStatus - -- | The response status code.

data DeregisterPatchBaselineForPatchGroupResponse

See: deregisterPatchBaselineForPatchGroupResponse smart constructor.

Response Lenses

dpbfpgrsBaselineId :: Lens' DeregisterPatchBaselineForPatchGroupResponse (Maybe Text)

The ID of the patch baseline the patch group was deregistered from.

dpbfpgrsPatchGroup :: Lens' DeregisterPatchBaselineForPatchGroupResponse (Maybe Text)

The name of the patch group deregistered from the patch baseline.

dpbfpgrsResponseStatus :: Lens' DeregisterPatchBaselineForPatchGroupResponse Int

- | The response status code.