Network.AWS.SSM.DescribeActivations

Description

Details about the activation, including: the date and time the activation was created, the expiration date, the IAM role assigned to the instances in the activation, and the number of instances activated by this registration.

This operation returns paginated results.

Creating a Request

describeActivations :: DescribeActivations

Creates a value of DescribeActivations with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

daFilters - A filter to view information about your activations.

daNextToken - A token to start the list. Use this token to get the next set of results.

daMaxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

data DescribeActivations

See: describeActivations smart constructor.

Request Lenses

daFilters :: Lens' DescribeActivations [DescribeActivationsFilter]

A filter to view information about your activations.

daNextToken :: Lens' DescribeActivations (Maybe Text)

A token to start the list. Use this token to get the next set of results.

daMaxResults :: Lens' DescribeActivations (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

Destructuring the Response

describeActivationsResponse

Arguments

:: Int
darsResponseStatus

-> DescribeActivationsResponse

Creates a value of DescribeActivationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

darsActivationList - A list of activations for your AWS account.

darsNextToken - The token for the next set of items to return. Use this token to get the next set of results.

darsResponseStatus - -- | The response status code.

data DescribeActivationsResponse

See: describeActivationsResponse smart constructor.

Response Lenses

darsActivationList :: Lens' DescribeActivationsResponse [Activation]

A list of activations for your AWS account.

darsNextToken :: Lens' DescribeActivationsResponse (Maybe Text)

The token for the next set of items to return. Use this token to get the next set of results.

darsResponseStatus :: Lens' DescribeActivationsResponse Int

- | The response status code.