Network.AWS.SSM.DescribeAvailablePatches

Description

Lists all patches that could possibly be included in a patch baseline.

Creating a Request

describeAvailablePatches :: DescribeAvailablePatches

Creates a value of DescribeAvailablePatches with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dapFilters - Filters used to scope down the returned patches.

dapNextToken - The token for the next set of items to return. (You received this token from a previous call.)

dapMaxResults - The maximum number of patches to return (per page).

data DescribeAvailablePatches

See: describeAvailablePatches smart constructor.

Request Lenses

dapFilters :: Lens' DescribeAvailablePatches [PatchOrchestratorFilter]

Filters used to scope down the returned patches.

dapNextToken :: Lens' DescribeAvailablePatches (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

dapMaxResults :: Lens' DescribeAvailablePatches (Maybe Natural)

The maximum number of patches to return (per page).

Destructuring the Response

describeAvailablePatchesResponse

Arguments

:: Int
daprsResponseStatus

-> DescribeAvailablePatchesResponse

Creates a value of DescribeAvailablePatchesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

daprsPatches - An array of patches. Each entry in the array is a patch structure.

daprsNextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

daprsResponseStatus - -- | The response status code.

data DescribeAvailablePatchesResponse

See: describeAvailablePatchesResponse smart constructor.

Response Lenses

daprsPatches :: Lens' DescribeAvailablePatchesResponse [Patch]

An array of patches. Each entry in the array is a patch structure.

daprsNextToken :: Lens' DescribeAvailablePatchesResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

daprsResponseStatus :: Lens' DescribeAvailablePatchesResponse Int

- | The response status code.