Network.AWS.SSM.DescribeDocument

Description

Describes the specified SSM document.

Creating a Request

describeDocument

Arguments

:: Text
ddName

-> DescribeDocument

Creates a value of DescribeDocument with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddDocumentVersion - The document version for which you want information. Can be a specific version or the default version.

ddName - The name of the SSM document.

data DescribeDocument

See: describeDocument smart constructor.

Request Lenses

ddDocumentVersion :: Lens' DescribeDocument (Maybe Text)

The document version for which you want information. Can be a specific version or the default version.

ddName :: Lens' DescribeDocument Text

The name of the SSM document.

Destructuring the Response

describeDocumentResponse

Arguments

:: Int
desrsResponseStatus

-> DescribeDocumentResponse

Creates a value of DescribeDocumentResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

desrsDocument - Information about the SSM document.

desrsResponseStatus - -- | The response status code.

data DescribeDocumentResponse

See: describeDocumentResponse smart constructor.

Response Lenses

desrsDocument :: Lens' DescribeDocumentResponse (Maybe DocumentDescription)

Information about the SSM document.

desrsResponseStatus :: Lens' DescribeDocumentResponse Int

- | The response status code.