Network.AWS.SSM.DescribeDocumentPermission

Description

Describes the permissions for a Systems Manager document. If you created the document, you are the owner. If a document is shared, it can either be shared privately (by specifying a user's AWS account ID) or publicly (All ).

Creating a Request

describeDocumentPermission

Arguments

:: Text
ddpName

-> DocumentPermissionType
ddpPermissionType

-> DescribeDocumentPermission

Creates a value of DescribeDocumentPermission with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddpName - The name of the document for which you are the owner.

ddpPermissionType - The permission type for the document. The permission type can be Share .

data DescribeDocumentPermission

See: describeDocumentPermission smart constructor.

Request Lenses

ddpName :: Lens' DescribeDocumentPermission Text

The name of the document for which you are the owner.

ddpPermissionType :: Lens' DescribeDocumentPermission DocumentPermissionType

The permission type for the document. The permission type can be Share .

Destructuring the Response

describeDocumentPermissionResponse

Arguments

:: Int
ddprsResponseStatus

-> DescribeDocumentPermissionResponse

Creates a value of DescribeDocumentPermissionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddprsAccountIds - The account IDs that have permission to use this document. The ID can be either an AWS account or All .

ddprsResponseStatus - -- | The response status code.

data DescribeDocumentPermissionResponse

See: describeDocumentPermissionResponse smart constructor.

Response Lenses

ddprsAccountIds :: Lens' DescribeDocumentPermissionResponse [Text]

The account IDs that have permission to use this document. The ID can be either an AWS account or All .

ddprsResponseStatus :: Lens' DescribeDocumentPermissionResponse Int

- | The response status code.