Network.AWS.SSM.DescribeInstanceAssociationsStatus

Description

The status of the associations for the instance(s).

Creating a Request

describeInstanceAssociationsStatus

Arguments

:: Text
diasInstanceId

-> DescribeInstanceAssociationsStatus

Creates a value of DescribeInstanceAssociationsStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diasNextToken - The token for the next set of items to return. (You received this token from a previous call.)

diasMaxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

diasInstanceId - The instance IDs for which you want association status information.

data DescribeInstanceAssociationsStatus

See: describeInstanceAssociationsStatus smart constructor.

Request Lenses

diasNextToken :: Lens' DescribeInstanceAssociationsStatus (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

diasMaxResults :: Lens' DescribeInstanceAssociationsStatus (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

diasInstanceId :: Lens' DescribeInstanceAssociationsStatus Text

The instance IDs for which you want association status information.

Destructuring the Response

describeInstanceAssociationsStatusResponse

Arguments

:: Int
diasrsResponseStatus

-> DescribeInstanceAssociationsStatusResponse

Creates a value of DescribeInstanceAssociationsStatusResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diasrsInstanceAssociationStatusInfos - Status information about the association.

diasrsNextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

diasrsResponseStatus - -- | The response status code.

data DescribeInstanceAssociationsStatusResponse

See: describeInstanceAssociationsStatusResponse smart constructor.

Response Lenses

diasrsInstanceAssociationStatusInfos :: Lens' DescribeInstanceAssociationsStatusResponse [InstanceAssociationStatusInfo]

Status information about the association.

diasrsNextToken :: Lens' DescribeInstanceAssociationsStatusResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

diasrsResponseStatus :: Lens' DescribeInstanceAssociationsStatusResponse Int

- | The response status code.