Network.AWS.SSM.DescribeInstanceInformation

Description

Describes one or more of your instances. You can use this to get information about instances like the operating system platform, the SSM Agent version (Linux), status etc. If you specify one or more instance IDs, it returns information for those instances. If you do not specify instance IDs, it returns information for all your instances. If you specify an instance ID that is not valid or an instance that you do not own, you receive an error.

This operation returns paginated results.

Creating a Request

describeInstanceInformation :: DescribeInstanceInformation

Creates a value of DescribeInstanceInformation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diiInstanceInformationFilterList - One or more filters. Use a filter to return a more specific list of instances.

diiFilters - One or more filters. Use a filter to return a more specific list of instances.

diiNextToken - The token for the next set of items to return. (You received this token from a previous call.)

diiMaxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

data DescribeInstanceInformation

See: describeInstanceInformation smart constructor.

Request Lenses

diiInstanceInformationFilterList :: Lens' DescribeInstanceInformation [InstanceInformationFilter]

One or more filters. Use a filter to return a more specific list of instances.

diiFilters :: Lens' DescribeInstanceInformation [InstanceInformationStringFilter]

One or more filters. Use a filter to return a more specific list of instances.

diiNextToken :: Lens' DescribeInstanceInformation (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

diiMaxResults :: Lens' DescribeInstanceInformation (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

Destructuring the Response

describeInstanceInformationResponse

Arguments

:: Int
diirsResponseStatus

-> DescribeInstanceInformationResponse

Creates a value of DescribeInstanceInformationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diirsNextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

diirsInstanceInformationList - The instance information list.

diirsResponseStatus - -- | The response status code.

data DescribeInstanceInformationResponse

See: describeInstanceInformationResponse smart constructor.

Response Lenses

diirsNextToken :: Lens' DescribeInstanceInformationResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

diirsInstanceInformationList :: Lens' DescribeInstanceInformationResponse [InstanceInformation]

The instance information list.

diirsResponseStatus :: Lens' DescribeInstanceInformationResponse Int

- | The response status code.