Network.AWS.SSM.DescribeInstancePatchStatesForPatchGroup

Description

Retrieves the high-level patch state for the instances in the specified patch group.

Creating a Request

describeInstancePatchStatesForPatchGroup

Arguments

:: Text
dipsfpgPatchGroup

-> DescribeInstancePatchStatesForPatchGroup

Creates a value of DescribeInstancePatchStatesForPatchGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dipsfpgFilters - Each entry in the array is a structure containing: Key (string between 1 and 200 characters) Values (array containing a single string) Type (string Equal, NotEqual, LessThan, GreaterThan)

dipsfpgNextToken - The token for the next set of items to return. (You received this token from a previous call.)

dipsfpgMaxResults - The maximum number of patches to return (per page).

dipsfpgPatchGroup - The name of the patch group for which the patch state information should be retrieved.

data DescribeInstancePatchStatesForPatchGroup

See: describeInstancePatchStatesForPatchGroup smart constructor.

Request Lenses

dipsfpgFilters :: Lens' DescribeInstancePatchStatesForPatchGroup [InstancePatchStateFilter]

Each entry in the array is a structure containing: Key (string between 1 and 200 characters) Values (array containing a single string) Type (string Equal, NotEqual, LessThan, GreaterThan)

dipsfpgNextToken :: Lens' DescribeInstancePatchStatesForPatchGroup (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

dipsfpgMaxResults :: Lens' DescribeInstancePatchStatesForPatchGroup (Maybe Natural)

The maximum number of patches to return (per page).

dipsfpgPatchGroup :: Lens' DescribeInstancePatchStatesForPatchGroup Text

The name of the patch group for which the patch state information should be retrieved.

Destructuring the Response

describeInstancePatchStatesForPatchGroupResponse

Arguments

:: Int
dipsfpgrsResponseStatus

-> DescribeInstancePatchStatesForPatchGroupResponse

Creates a value of DescribeInstancePatchStatesForPatchGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dipsfpgrsNextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dipsfpgrsInstancePatchStates - The high-level patch state for the requested instances.

dipsfpgrsResponseStatus - -- | The response status code.

data DescribeInstancePatchStatesForPatchGroupResponse

See: describeInstancePatchStatesForPatchGroupResponse smart constructor.

Response Lenses

dipsfpgrsNextToken :: Lens' DescribeInstancePatchStatesForPatchGroupResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dipsfpgrsInstancePatchStates :: Lens' DescribeInstancePatchStatesForPatchGroupResponse (Maybe (NonEmpty InstancePatchState))

The high-level patch state for the requested instances.

dipsfpgrsResponseStatus :: Lens' DescribeInstancePatchStatesForPatchGroupResponse Int

- | The response status code.