Network.AWS.SSM.DescribeInstancePatches

Description

Retrieves information about the patches on the specified instance and their state relative to the patch baseline being used for the instance.

Creating a Request

describeInstancePatches

Arguments

:: Text
dipInstanceId

-> DescribeInstancePatches

Creates a value of DescribeInstancePatches with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dipFilters - Each entry in the array is a structure containing: Key (string, between 1 and 128 characters) Values (array of strings, each string between 1 and 256 characters)

dipNextToken - The token for the next set of items to return. (You received this token from a previous call.)

dipMaxResults - The maximum number of patches to return (per page).

dipInstanceId - The ID of the instance whose patch state information should be retrieved.

data DescribeInstancePatches

See: describeInstancePatches smart constructor.

Request Lenses

dipFilters :: Lens' DescribeInstancePatches [PatchOrchestratorFilter]

Each entry in the array is a structure containing: Key (string, between 1 and 128 characters) Values (array of strings, each string between 1 and 256 characters)

dipNextToken :: Lens' DescribeInstancePatches (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

dipMaxResults :: Lens' DescribeInstancePatches (Maybe Natural)

The maximum number of patches to return (per page).

dipInstanceId :: Lens' DescribeInstancePatches Text

The ID of the instance whose patch state information should be retrieved.

Destructuring the Response

describeInstancePatchesResponse

Arguments

:: Int
diprsResponseStatus

-> DescribeInstancePatchesResponse

Creates a value of DescribeInstancePatchesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diprsPatches - Each entry in the array is a structure containing: Title (string) KBId (string) Classification (string) Severity (string) State (string: INSTALLED, "INSTALLED OTHER", MISSING, "NOT APPLICABLE", FAILED) InstalledTime (DateTime) InstalledBy (string)

diprsNextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

diprsResponseStatus - -- | The response status code.

data DescribeInstancePatchesResponse

See: describeInstancePatchesResponse smart constructor.

Response Lenses

diprsPatches :: Lens' DescribeInstancePatchesResponse [PatchComplianceData]

Each entry in the array is a structure containing: Title (string) KBId (string) Classification (string) Severity (string) State (string: INSTALLED, "INSTALLED OTHER", MISSING, "NOT APPLICABLE", FAILED) InstalledTime (DateTime) InstalledBy (string)

diprsNextToken :: Lens' DescribeInstancePatchesResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

diprsResponseStatus :: Lens' DescribeInstancePatchesResponse Int

- | The response status code.