Network.AWS.SSM.DescribeMaintenanceWindowExecutionTaskInvocations

Description

Retrieves the individual task executions (one per target) for a particular task executed as part of a Maintenance Window execution.

Creating a Request

describeMaintenanceWindowExecutionTaskInvocations

Arguments

:: Text
dmwetiWindowExecutionId

-> Text
dmwetiTaskId

-> DescribeMaintenanceWindowExecutionTaskInvocations

Creates a value of DescribeMaintenanceWindowExecutionTaskInvocations with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmwetiFilters - Optional filters used to scope down the returned task invocations. The supported filter key is STATUS with the corresponding values PENDING, IN_PROGRESS, SUCCESS, FAILED, TIMED_OUT, CANCELLING, and CANCELLED.

dmwetiNextToken - The token for the next set of items to return. (You received this token from a previous call.)

dmwetiMaxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

dmwetiWindowExecutionId - The ID of the Maintenance Window execution the task is part of.

dmwetiTaskId - The ID of the specific task in the Maintenance Window task that should be retrieved.

data DescribeMaintenanceWindowExecutionTaskInvocations

See: describeMaintenanceWindowExecutionTaskInvocations smart constructor.

Request Lenses

dmwetiFilters :: Lens' DescribeMaintenanceWindowExecutionTaskInvocations [MaintenanceWindowFilter]

Optional filters used to scope down the returned task invocations. The supported filter key is STATUS with the corresponding values PENDING, IN_PROGRESS, SUCCESS, FAILED, TIMED_OUT, CANCELLING, and CANCELLED.

dmwetiNextToken :: Lens' DescribeMaintenanceWindowExecutionTaskInvocations (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

dmwetiMaxResults :: Lens' DescribeMaintenanceWindowExecutionTaskInvocations (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

dmwetiWindowExecutionId :: Lens' DescribeMaintenanceWindowExecutionTaskInvocations Text

The ID of the Maintenance Window execution the task is part of.

dmwetiTaskId :: Lens' DescribeMaintenanceWindowExecutionTaskInvocations Text

The ID of the specific task in the Maintenance Window task that should be retrieved.

Destructuring the Response

describeMaintenanceWindowExecutionTaskInvocationsResponse

Arguments

:: Int
dmwetirsResponseStatus

-> DescribeMaintenanceWindowExecutionTaskInvocationsResponse

Creates a value of DescribeMaintenanceWindowExecutionTaskInvocationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmwetirsWindowExecutionTaskInvocationIdentities - Information about the task invocation results per invocation.

dmwetirsNextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dmwetirsResponseStatus - -- | The response status code.

data DescribeMaintenanceWindowExecutionTaskInvocationsResponse

See: describeMaintenanceWindowExecutionTaskInvocationsResponse smart constructor.

Response Lenses

dmwetirsWindowExecutionTaskInvocationIdentities :: Lens' DescribeMaintenanceWindowExecutionTaskInvocationsResponse [MaintenanceWindowExecutionTaskInvocationIdentity]

Information about the task invocation results per invocation.

dmwetirsNextToken :: Lens' DescribeMaintenanceWindowExecutionTaskInvocationsResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dmwetirsResponseStatus :: Lens' DescribeMaintenanceWindowExecutionTaskInvocationsResponse Int

- | The response status code.