Network.AWS.SSM.DescribeMaintenanceWindowExecutions

Description

Lists the executions of a Maintenance Window. This includes information about when the Maintenance Window was scheduled to be active, and information about tasks registered and run with the Maintenance Window.

Creating a Request

describeMaintenanceWindowExecutions

Arguments

:: Text
dmweWindowId

-> DescribeMaintenanceWindowExecutions

Creates a value of DescribeMaintenanceWindowExecutions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmweFilters - Each entry in the array is a structure containing: Key (string, between 1 and 128 characters) Values (array of strings, each string is between 1 and 256 characters) The supported Keys are ExecutedBefore and ExecutedAfter with the value being a date/time string such as 2016-11-04T05:00:00Z.

dmweNextToken - The token for the next set of items to return. (You received this token from a previous call.)

dmweMaxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

dmweWindowId - The ID of the Maintenance Window whose executions should be retrieved.

data DescribeMaintenanceWindowExecutions

See: describeMaintenanceWindowExecutions smart constructor.

Request Lenses

dmweFilters :: Lens' DescribeMaintenanceWindowExecutions [MaintenanceWindowFilter]

Each entry in the array is a structure containing: Key (string, between 1 and 128 characters) Values (array of strings, each string is between 1 and 256 characters) The supported Keys are ExecutedBefore and ExecutedAfter with the value being a date/time string such as 2016-11-04T05:00:00Z.

dmweNextToken :: Lens' DescribeMaintenanceWindowExecutions (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

dmweMaxResults :: Lens' DescribeMaintenanceWindowExecutions (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

dmweWindowId :: Lens' DescribeMaintenanceWindowExecutions Text

The ID of the Maintenance Window whose executions should be retrieved.

Destructuring the Response

describeMaintenanceWindowExecutionsResponse

Arguments

:: Int
dmwersResponseStatus

-> DescribeMaintenanceWindowExecutionsResponse

Creates a value of DescribeMaintenanceWindowExecutionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmwersWindowExecutions - Information about the Maintenance Windows execution.

dmwersNextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dmwersResponseStatus - -- | The response status code.

data DescribeMaintenanceWindowExecutionsResponse

See: describeMaintenanceWindowExecutionsResponse smart constructor.

Response Lenses

dmwersWindowExecutions :: Lens' DescribeMaintenanceWindowExecutionsResponse [MaintenanceWindowExecution]

Information about the Maintenance Windows execution.

dmwersNextToken :: Lens' DescribeMaintenanceWindowExecutionsResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dmwersResponseStatus :: Lens' DescribeMaintenanceWindowExecutionsResponse Int

- | The response status code.