Network.AWS.SSM.DescribeMaintenanceWindowTargets

Description

Lists the targets registered with the Maintenance Window.

Creating a Request

describeMaintenanceWindowTargets

Arguments

:: Text
dmwtWindowId

-> DescribeMaintenanceWindowTargets

Creates a value of DescribeMaintenanceWindowTargets with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmwtFilters - Optional filters that can be used to narrow down the scope of the returned window targets. The supported filter keys are Type, WindowTargetId and OwnerInformation.

dmwtNextToken - The token for the next set of items to return. (You received this token from a previous call.)

dmwtMaxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

dmwtWindowId - The ID of the Maintenance Window whose targets should be retrieved.

data DescribeMaintenanceWindowTargets

See: describeMaintenanceWindowTargets smart constructor.

Request Lenses

dmwtFilters :: Lens' DescribeMaintenanceWindowTargets [MaintenanceWindowFilter]

Optional filters that can be used to narrow down the scope of the returned window targets. The supported filter keys are Type, WindowTargetId and OwnerInformation.

dmwtNextToken :: Lens' DescribeMaintenanceWindowTargets (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

dmwtMaxResults :: Lens' DescribeMaintenanceWindowTargets (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

dmwtWindowId :: Lens' DescribeMaintenanceWindowTargets Text

The ID of the Maintenance Window whose targets should be retrieved.

Destructuring the Response

describeMaintenanceWindowTargetsResponse

Arguments

:: Int
dmwtrsResponseStatus

-> DescribeMaintenanceWindowTargetsResponse

Creates a value of DescribeMaintenanceWindowTargetsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmwtrsNextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dmwtrsTargets - Information about the targets in the Maintenance Window.

dmwtrsResponseStatus - -- | The response status code.

data DescribeMaintenanceWindowTargetsResponse

See: describeMaintenanceWindowTargetsResponse smart constructor.

Response Lenses

dmwtrsNextToken :: Lens' DescribeMaintenanceWindowTargetsResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dmwtrsTargets :: Lens' DescribeMaintenanceWindowTargetsResponse [MaintenanceWindowTarget]

Information about the targets in the Maintenance Window.

dmwtrsResponseStatus :: Lens' DescribeMaintenanceWindowTargetsResponse Int

- | The response status code.