Network.AWS.SSM.DescribeParameters

Description

Get information about a parameter.

Request results are returned on a best-effort basis. If you specify MaxResults in the request, the response includes information up to the limit specified. The number of items returned, however, can be between zero and the value of MaxResults . If the service reaches an internal limit while processing the results, it stops the operation and returns the matching values up to that point and a NextToken . You can specify the NextToken in a subsequent call to get the next set of results.

This operation returns paginated results.

Creating a Request

describeParameters :: DescribeParameters

Creates a value of DescribeParameters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dpParameterFilters - Filters to limit the request results.

dpFilters - One or more filters. Use a filter to return a more specific list of results.

dpNextToken - The token for the next set of items to return. (You received this token from a previous call.)

dpMaxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

data DescribeParameters

See: describeParameters smart constructor.

Request Lenses

dpParameterFilters :: Lens' DescribeParameters [ParameterStringFilter]

Filters to limit the request results.

dpFilters :: Lens' DescribeParameters [ParametersFilter]

One or more filters. Use a filter to return a more specific list of results.

dpNextToken :: Lens' DescribeParameters (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

dpMaxResults :: Lens' DescribeParameters (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

Destructuring the Response

describeParametersResponse

Arguments

:: Int
dprsResponseStatus

-> DescribeParametersResponse

Creates a value of DescribeParametersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dprsNextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dprsParameters - Parameters returned by the request.

dprsResponseStatus - -- | The response status code.

data DescribeParametersResponse

See: describeParametersResponse smart constructor.

Response Lenses

dprsNextToken :: Lens' DescribeParametersResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dprsParameters :: Lens' DescribeParametersResponse [ParameterMetadata]

Parameters returned by the request.

dprsResponseStatus :: Lens' DescribeParametersResponse Int

- | The response status code.