Network.AWS.SSM.GetAutomationExecution

Description

Get detailed information about a particular Automation execution.

Creating a Request

getAutomationExecution

Arguments

:: Text
gaeAutomationExecutionId

-> GetAutomationExecution

Creates a value of GetAutomationExecution with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gaeAutomationExecutionId - The unique identifier for an existing automation execution to examine. The execution ID is returned by StartAutomationExecution when the execution of an Automation document is initiated.

data GetAutomationExecution

See: getAutomationExecution smart constructor.

Request Lenses

gaeAutomationExecutionId :: Lens' GetAutomationExecution Text

The unique identifier for an existing automation execution to examine. The execution ID is returned by StartAutomationExecution when the execution of an Automation document is initiated.

Destructuring the Response

getAutomationExecutionResponse

Arguments

:: Int
gaersResponseStatus

-> GetAutomationExecutionResponse

Creates a value of GetAutomationExecutionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gaersAutomationExecution - Detailed information about the current state of an automation execution.

gaersResponseStatus - -- | The response status code.

data GetAutomationExecutionResponse

See: getAutomationExecutionResponse smart constructor.

Response Lenses

gaersAutomationExecution :: Lens' GetAutomationExecutionResponse (Maybe AutomationExecution)

Detailed information about the current state of an automation execution.

gaersResponseStatus :: Lens' GetAutomationExecutionResponse Int

- | The response status code.