Network.AWS.SSM.GetCommandInvocation

Description

Returns detailed information about command execution for an invocation or plugin.

Creating a Request

getCommandInvocation

Arguments

:: Text
gciCommandId

-> Text
gciInstanceId

-> GetCommandInvocation

Creates a value of GetCommandInvocation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gciPluginName - (Optional) The name of the plugin for which you want detailed results. If the document contains only one plugin, the name can be omitted and the details will be returned.

gciCommandId - (Required) The parent command ID of the invocation plugin.

gciInstanceId - (Required) The ID of the managed instance targeted by the command. A managed instance can be an Amazon EC2 instance or an instance in your hybrid environment that is configured for Systems Manager.

data GetCommandInvocation

See: getCommandInvocation smart constructor.

Request Lenses

gciPluginName :: Lens' GetCommandInvocation (Maybe Text)

(Optional) The name of the plugin for which you want detailed results. If the document contains only one plugin, the name can be omitted and the details will be returned.

gciCommandId :: Lens' GetCommandInvocation Text

(Required) The parent command ID of the invocation plugin.

gciInstanceId :: Lens' GetCommandInvocation Text

(Required) The ID of the managed instance targeted by the command. A managed instance can be an Amazon EC2 instance or an instance in your hybrid environment that is configured for Systems Manager.

Destructuring the Response

getCommandInvocationResponse

Arguments

:: Int
gcirsResponseStatus

-> GetCommandInvocationResponse

Creates a value of GetCommandInvocationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcirsInstanceId - The ID of the managed instance targeted by the command. A managed instance can be an Amazon EC2 instance or an instance in your hybrid environment that is configured for Systems Manager.

gcirsStatus - The status of the parent command for this invocation. This status can be different than StatusDetails.

gcirsStandardErrorContent - The first 8,000 characters written by the plugin to stderr. If the command has not finished executing, then this string is empty.

gcirsExecutionElapsedTime - Duration since ExecutionStartDateTime.

gcirsDocumentName - The name of the document that was executed. For example, AWS-RunShellScript.

gcirsStandardErrorURL - The URL for the complete text written by the plugin to stderr. If the command has not finished executing, then this string is empty.

gcirsExecutionStartDateTime - The date and time the plugin started executing. Date and time are written in ISO 8601 format. For example, June 7, 2017 is represented as 2017-06-7. The following sample AWS CLI command uses the InvokedBefore filter. aws ssm list-commands --filters key=InvokedBefore,value=2017-06-07T00:00:00Z If the plugin has not started to execute, the string is empty.

gcirsResponseCode - The error level response code for the plugin script. If the response code is -1, then the command has not started executing on the instance, or it was not received by the instance.

gcirsStatusDetails - A detailed status of the command execution for an invocation. StatusDetails includes more information than Status because it includes states resulting from error and concurrency control parameters. StatusDetails can show different results than Status. For more information about these statuses, see Run Command Status . StatusDetails can be one of the following values: * Pending: The command has not been sent to the instance. * In Progress: The command has been sent to the instance but has not reached a terminal state. * Delayed: The system attempted to send the command to the target, but the target was not available. The instance might not be available because of network issues, the instance was stopped, etc. The system will try to deliver the command again. * Success: The command or plugin was executed successfully. This is a terminal state. * Delivery Timed Out: The command was not delivered to the instance before the delivery timeout expired. Delivery timeouts do not count against the parent command's MaxErrors limit, but they do contribute to whether the parent command status is Success or Incomplete. This is a terminal state. * Execution Timed Out: The command started to execute on the instance, but the execution was not complete before the timeout expired. Execution timeouts count against the MaxErrors limit of the parent command. This is a terminal state. * Failed: The command wasn't executed successfully on the instance. For a plugin, this indicates that the result code was not zero. For a command invocation, this indicates that the result code for one or more plugins was not zero. Invocation failures count against the MaxErrors limit of the parent command. This is a terminal state. * Canceled: The command was terminated before it was completed. This is a terminal state. * Undeliverable: The command can't be delivered to the instance. The instance might not exist or might not be responding. Undeliverable invocations don't count against the parent command's MaxErrors limit and don't contribute to whether the parent command status is Success or Incomplete. This is a terminal state. * Terminated: The parent command exceeded its MaxErrors limit and subsequent command invocations were canceled by the system. This is a terminal state.

gcirsExecutionEndDateTime - The date and time the plugin was finished executing. Date and time are written in ISO 8601 format. For example, June 7, 2017 is represented as 2017-06-7. The following sample AWS CLI command uses the InvokedAfter filter. aws ssm list-commands --filters key=InvokedAfter,value=2017-06-07T00:00:00Z If the plugin has not started to execute, the string is empty.

gcirsStandardOutputURL - The URL for the complete text written by the plugin to stdout in Amazon S3. If an Amazon S3 bucket was not specified, then this string is empty.

gcirsCommandId - The parent command ID of the invocation plugin.

gcirsStandardOutputContent - The first 24,000 characters written by the plugin to stdout. If the command has not finished executing, if ExecutionStatus is neither Succeeded nor Failed, then this string is empty.

gcirsComment - The comment text for the command.

gcirsPluginName - The name of the plugin for which you want detailed results. For example, aws:RunShellScript is a plugin.

gcirsResponseStatus - -- | The response status code.

data GetCommandInvocationResponse

See: getCommandInvocationResponse smart constructor.

Response Lenses

gcirsInstanceId :: Lens' GetCommandInvocationResponse (Maybe Text)

The ID of the managed instance targeted by the command. A managed instance can be an Amazon EC2 instance or an instance in your hybrid environment that is configured for Systems Manager.

gcirsStatus :: Lens' GetCommandInvocationResponse (Maybe CommandInvocationStatus)

The status of the parent command for this invocation. This status can be different than StatusDetails.

gcirsStandardErrorContent :: Lens' GetCommandInvocationResponse (Maybe Text)

The first 8,000 characters written by the plugin to stderr. If the command has not finished executing, then this string is empty.

gcirsExecutionElapsedTime :: Lens' GetCommandInvocationResponse (Maybe Text)

Duration since ExecutionStartDateTime.

gcirsDocumentName :: Lens' GetCommandInvocationResponse (Maybe Text)

The name of the document that was executed. For example, AWS-RunShellScript.

gcirsStandardErrorURL :: Lens' GetCommandInvocationResponse (Maybe Text)

The URL for the complete text written by the plugin to stderr. If the command has not finished executing, then this string is empty.

gcirsExecutionStartDateTime :: Lens' GetCommandInvocationResponse (Maybe Text)

The date and time the plugin started executing. Date and time are written in ISO 8601 format. For example, June 7, 2017 is represented as 2017-06-7. The following sample AWS CLI command uses the InvokedBefore filter. aws ssm list-commands --filters key=InvokedBefore,value=2017-06-07T00:00:00Z If the plugin has not started to execute, the string is empty.

gcirsResponseCode :: Lens' GetCommandInvocationResponse (Maybe Int)

The error level response code for the plugin script. If the response code is -1, then the command has not started executing on the instance, or it was not received by the instance.

gcirsStatusDetails :: Lens' GetCommandInvocationResponse (Maybe Text)

A detailed status of the command execution for an invocation. StatusDetails includes more information than Status because it includes states resulting from error and concurrency control parameters. StatusDetails can show different results than Status. For more information about these statuses, see Run Command Status . StatusDetails can be one of the following values: * Pending: The command has not been sent to the instance. * In Progress: The command has been sent to the instance but has not reached a terminal state. * Delayed: The system attempted to send the command to the target, but the target was not available. The instance might not be available because of network issues, the instance was stopped, etc. The system will try to deliver the command again. * Success: The command or plugin was executed successfully. This is a terminal state. * Delivery Timed Out: The command was not delivered to the instance before the delivery timeout expired. Delivery timeouts do not count against the parent command's MaxErrors limit, but they do contribute to whether the parent command status is Success or Incomplete. This is a terminal state. * Execution Timed Out: The command started to execute on the instance, but the execution was not complete before the timeout expired. Execution timeouts count against the MaxErrors limit of the parent command. This is a terminal state. * Failed: The command wasn't executed successfully on the instance. For a plugin, this indicates that the result code was not zero. For a command invocation, this indicates that the result code for one or more plugins was not zero. Invocation failures count against the MaxErrors limit of the parent command. This is a terminal state. * Canceled: The command was terminated before it was completed. This is a terminal state. * Undeliverable: The command can't be delivered to the instance. The instance might not exist or might not be responding. Undeliverable invocations don't count against the parent command's MaxErrors limit and don't contribute to whether the parent command status is Success or Incomplete. This is a terminal state. * Terminated: The parent command exceeded its MaxErrors limit and subsequent command invocations were canceled by the system. This is a terminal state.

gcirsExecutionEndDateTime :: Lens' GetCommandInvocationResponse (Maybe Text)

The date and time the plugin was finished executing. Date and time are written in ISO 8601 format. For example, June 7, 2017 is represented as 2017-06-7. The following sample AWS CLI command uses the InvokedAfter filter. aws ssm list-commands --filters key=InvokedAfter,value=2017-06-07T00:00:00Z If the plugin has not started to execute, the string is empty.

gcirsStandardOutputURL :: Lens' GetCommandInvocationResponse (Maybe Text)

The URL for the complete text written by the plugin to stdout in Amazon S3. If an Amazon S3 bucket was not specified, then this string is empty.

gcirsCommandId :: Lens' GetCommandInvocationResponse (Maybe Text)

The parent command ID of the invocation plugin.

gcirsStandardOutputContent :: Lens' GetCommandInvocationResponse (Maybe Text)

The first 24,000 characters written by the plugin to stdout. If the command has not finished executing, if ExecutionStatus is neither Succeeded nor Failed, then this string is empty.

gcirsComment :: Lens' GetCommandInvocationResponse (Maybe Text)

The comment text for the command.

gcirsPluginName :: Lens' GetCommandInvocationResponse (Maybe Text)

The name of the plugin for which you want detailed results. For example, aws:RunShellScript is a plugin.

gcirsResponseStatus :: Lens' GetCommandInvocationResponse Int

- | The response status code.