Network.AWS.SSM.GetDefaultPatchBaseline

Description

Retrieves the default patch baseline. Note that Systems Manager supports creating multiple default patch baselines. For example, you can create a default patch baseline for each operating system.

Creating a Request

getDefaultPatchBaseline :: GetDefaultPatchBaseline

Creates a value of GetDefaultPatchBaseline with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdpbOperatingSystem - Returns the default patch baseline for the specified operating system.

data GetDefaultPatchBaseline

See: getDefaultPatchBaseline smart constructor.

Request Lenses

gdpbOperatingSystem :: Lens' GetDefaultPatchBaseline (Maybe OperatingSystem)

Returns the default patch baseline for the specified operating system.

Destructuring the Response

getDefaultPatchBaselineResponse

Arguments

:: Int
gdpbrsResponseStatus

-> GetDefaultPatchBaselineResponse

Creates a value of GetDefaultPatchBaselineResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdpbrsOperatingSystem - The operating system for the returned patch baseline.

gdpbrsBaselineId - The ID of the default patch baseline.

gdpbrsResponseStatus - -- | The response status code.

data GetDefaultPatchBaselineResponse

See: getDefaultPatchBaselineResponse smart constructor.

Response Lenses

gdpbrsOperatingSystem :: Lens' GetDefaultPatchBaselineResponse (Maybe OperatingSystem)

The operating system for the returned patch baseline.

gdpbrsBaselineId :: Lens' GetDefaultPatchBaselineResponse (Maybe Text)

The ID of the default patch baseline.

gdpbrsResponseStatus :: Lens' GetDefaultPatchBaselineResponse Int

- | The response status code.