Network.AWS.SSM.GetDeployablePatchSnapshotForInstance

Description

Retrieves the current snapshot for the patch baseline the instance uses. This API is primarily used by the AWS-RunPatchBaseline Systems Manager document.

Creating a Request

getDeployablePatchSnapshotForInstance

Arguments

:: Text
gdpsfiInstanceId

-> Text
gdpsfiSnapshotId

-> GetDeployablePatchSnapshotForInstance

Creates a value of GetDeployablePatchSnapshotForInstance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdpsfiInstanceId - The ID of the instance for which the appropriate patch snapshot should be retrieved.

gdpsfiSnapshotId - The user-defined snapshot ID.

data GetDeployablePatchSnapshotForInstance

See: getDeployablePatchSnapshotForInstance smart constructor.

Request Lenses

gdpsfiInstanceId :: Lens' GetDeployablePatchSnapshotForInstance Text

The ID of the instance for which the appropriate patch snapshot should be retrieved.

gdpsfiSnapshotId :: Lens' GetDeployablePatchSnapshotForInstance Text

The user-defined snapshot ID.

Destructuring the Response

getDeployablePatchSnapshotForInstanceResponse

Arguments

:: Int
gdpsfirsResponseStatus

-> GetDeployablePatchSnapshotForInstanceResponse

Creates a value of GetDeployablePatchSnapshotForInstanceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdpsfirsInstanceId - The ID of the instance.

gdpsfirsProduct - Returns the specific operating system (for example Windows Server 2012 or Amazon Linux 2015.09) on the instance for the specified patch snapshot.

gdpsfirsSnapshotDownloadURL - A pre-signed Amazon S3 URL that can be used to download the patch snapshot.

gdpsfirsSnapshotId - The user-defined snapshot ID.

gdpsfirsResponseStatus - -- | The response status code.

data GetDeployablePatchSnapshotForInstanceResponse

See: getDeployablePatchSnapshotForInstanceResponse smart constructor.

Response Lenses

gdpsfirsInstanceId :: Lens' GetDeployablePatchSnapshotForInstanceResponse (Maybe Text)

The ID of the instance.

gdpsfirsProduct :: Lens' GetDeployablePatchSnapshotForInstanceResponse (Maybe Text)

Returns the specific operating system (for example Windows Server 2012 or Amazon Linux 2015.09) on the instance for the specified patch snapshot.

gdpsfirsSnapshotDownloadURL :: Lens' GetDeployablePatchSnapshotForInstanceResponse (Maybe Text)

A pre-signed Amazon S3 URL that can be used to download the patch snapshot.

gdpsfirsSnapshotId :: Lens' GetDeployablePatchSnapshotForInstanceResponse (Maybe Text)

The user-defined snapshot ID.

gdpsfirsResponseStatus :: Lens' GetDeployablePatchSnapshotForInstanceResponse Int

- | The response status code.