Network.AWS.SSM.GetDocument

Description

Gets the contents of the specified SSM document.

Creating a Request

getDocument

Arguments

:: Text
gdName

-> GetDocument

Creates a value of GetDocument with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdDocumentVersion - The document version for which you want information.

gdName - The name of the SSM document.

data GetDocument

See: getDocument smart constructor.

Request Lenses

gdDocumentVersion :: Lens' GetDocument (Maybe Text)

The document version for which you want information.

gdName :: Lens' GetDocument Text

The name of the SSM document.

Destructuring the Response

getDocumentResponse

Arguments

:: Int
gdrsResponseStatus

-> GetDocumentResponse

Creates a value of GetDocumentResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdrsDocumentType - The document type.

gdrsContent - The contents of the SSM document.

gdrsName - The name of the SSM document.

gdrsDocumentVersion - The document version.

gdrsResponseStatus - -- | The response status code.

data GetDocumentResponse

See: getDocumentResponse smart constructor.

Response Lenses

gdrsDocumentType :: Lens' GetDocumentResponse (Maybe DocumentType)

The document type.

gdrsContent :: Lens' GetDocumentResponse (Maybe Text)

The contents of the SSM document.

gdrsName :: Lens' GetDocumentResponse (Maybe Text)

The name of the SSM document.

gdrsDocumentVersion :: Lens' GetDocumentResponse (Maybe Text)

The document version.

gdrsResponseStatus :: Lens' GetDocumentResponse Int

- | The response status code.