Network.AWS.SSM.GetInventorySchema

Description

Return a list of inventory type names for the account, or return a list of attribute names for a specific Inventory item type.

Creating a Request

getInventorySchema :: GetInventorySchema

Creates a value of GetInventorySchema with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gisTypeName - The type of inventory item to return.

gisNextToken - The token for the next set of items to return. (You received this token from a previous call.)

gisSubType - Returns the sub-type schema for a specified inventory type.

gisMaxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

data GetInventorySchema

See: getInventorySchema smart constructor.

Request Lenses

gisTypeName :: Lens' GetInventorySchema (Maybe Text)

The type of inventory item to return.

gisNextToken :: Lens' GetInventorySchema (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

gisSubType :: Lens' GetInventorySchema (Maybe Bool)

Returns the sub-type schema for a specified inventory type.

gisMaxResults :: Lens' GetInventorySchema (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

Destructuring the Response

getInventorySchemaResponse

Arguments

:: Int
gisrsResponseStatus

-> GetInventorySchemaResponse

Creates a value of GetInventorySchemaResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gisrsSchemas - Inventory schemas returned by the request.

gisrsNextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

gisrsResponseStatus - -- | The response status code.

data GetInventorySchemaResponse

See: getInventorySchemaResponse smart constructor.

Response Lenses

gisrsSchemas :: Lens' GetInventorySchemaResponse [InventoryItemSchema]

Inventory schemas returned by the request.

gisrsNextToken :: Lens' GetInventorySchemaResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

gisrsResponseStatus :: Lens' GetInventorySchemaResponse Int

- | The response status code.