Network.AWS.SSM.GetMaintenanceWindow

Description

Retrieves a Maintenance Window.

Creating a Request

getMaintenanceWindow

Arguments

:: Text
gmwWindowId

-> GetMaintenanceWindow

Creates a value of GetMaintenanceWindow with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gmwWindowId - The ID of the desired Maintenance Window.

data GetMaintenanceWindow

See: getMaintenanceWindow smart constructor.

Request Lenses

gmwWindowId :: Lens' GetMaintenanceWindow Text

The ID of the desired Maintenance Window.

Destructuring the Response

getMaintenanceWindowResponse

Arguments

:: Int
gmwrsResponseStatus

-> GetMaintenanceWindowResponse

Creates a value of GetMaintenanceWindowResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gmwrsEnabled - Whether the Maintenance Windows is enabled.

gmwrsSchedule - The schedule of the Maintenance Window in the form of a cron or rate expression.

gmwrsCreatedDate - The date the Maintenance Window was created.

gmwrsName - The name of the Maintenance Window.

gmwrsModifiedDate - The date the Maintenance Window was last modified.

gmwrsCutoff - The number of hours before the end of the Maintenance Window that Systems Manager stops scheduling new tasks for execution.

gmwrsAllowUnassociatedTargets - Whether targets must be registered with the Maintenance Window before tasks can be defined for those targets.

gmwrsDescription - The description of the Maintenance Window.

gmwrsDuration - The duration of the Maintenance Window in hours.

gmwrsWindowId - The ID of the created Maintenance Window.

gmwrsResponseStatus - -- | The response status code.

data GetMaintenanceWindowResponse

See: getMaintenanceWindowResponse smart constructor.

Response Lenses

gmwrsEnabled :: Lens' GetMaintenanceWindowResponse (Maybe Bool)

Whether the Maintenance Windows is enabled.

gmwrsSchedule :: Lens' GetMaintenanceWindowResponse (Maybe Text)

The schedule of the Maintenance Window in the form of a cron or rate expression.

gmwrsCreatedDate :: Lens' GetMaintenanceWindowResponse (Maybe UTCTime)

The date the Maintenance Window was created.

gmwrsName :: Lens' GetMaintenanceWindowResponse (Maybe Text)

The name of the Maintenance Window.

gmwrsModifiedDate :: Lens' GetMaintenanceWindowResponse (Maybe UTCTime)

The date the Maintenance Window was last modified.

gmwrsCutoff :: Lens' GetMaintenanceWindowResponse (Maybe Natural)

The number of hours before the end of the Maintenance Window that Systems Manager stops scheduling new tasks for execution.

gmwrsAllowUnassociatedTargets :: Lens' GetMaintenanceWindowResponse (Maybe Bool)

Whether targets must be registered with the Maintenance Window before tasks can be defined for those targets.

gmwrsDescription :: Lens' GetMaintenanceWindowResponse (Maybe Text)

The description of the Maintenance Window.

gmwrsDuration :: Lens' GetMaintenanceWindowResponse (Maybe Natural)

The duration of the Maintenance Window in hours.

gmwrsWindowId :: Lens' GetMaintenanceWindowResponse (Maybe Text)

The ID of the created Maintenance Window.

gmwrsResponseStatus :: Lens' GetMaintenanceWindowResponse Int

- | The response status code.