Network.AWS.SSM.GetMaintenanceWindowExecution

Description

Retrieves details about a specific task executed as part of a Maintenance Window execution.

Creating a Request

getMaintenanceWindowExecution

Arguments

:: Text
gmweWindowExecutionId

-> GetMaintenanceWindowExecution

Creates a value of GetMaintenanceWindowExecution with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gmweWindowExecutionId - The ID of the Maintenance Window execution that includes the task.

data GetMaintenanceWindowExecution

See: getMaintenanceWindowExecution smart constructor.

Request Lenses

gmweWindowExecutionId :: Lens' GetMaintenanceWindowExecution Text

The ID of the Maintenance Window execution that includes the task.

Destructuring the Response

getMaintenanceWindowExecutionResponse

Arguments

:: Int
gmwersResponseStatus

-> GetMaintenanceWindowExecutionResponse

Creates a value of GetMaintenanceWindowExecutionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gmwersStatus - The status of the Maintenance Window execution.

gmwersStartTime - The time the Maintenance Window started executing.

gmwersWindowExecutionId - The ID of the Maintenance Window execution.

gmwersStatusDetails - The details explaining the Status. Only available for certain status values.

gmwersEndTime - The time the Maintenance Window finished executing.

gmwersTaskIds - The ID of the task executions from the Maintenance Window execution.

gmwersResponseStatus - -- | The response status code.

data GetMaintenanceWindowExecutionResponse

See: getMaintenanceWindowExecutionResponse smart constructor.

Response Lenses

gmwersStatus :: Lens' GetMaintenanceWindowExecutionResponse (Maybe MaintenanceWindowExecutionStatus)

The status of the Maintenance Window execution.

gmwersStartTime :: Lens' GetMaintenanceWindowExecutionResponse (Maybe UTCTime)

The time the Maintenance Window started executing.

gmwersWindowExecutionId :: Lens' GetMaintenanceWindowExecutionResponse (Maybe Text)

The ID of the Maintenance Window execution.

gmwersStatusDetails :: Lens' GetMaintenanceWindowExecutionResponse (Maybe Text)

The details explaining the Status. Only available for certain status values.

gmwersEndTime :: Lens' GetMaintenanceWindowExecutionResponse (Maybe UTCTime)

The time the Maintenance Window finished executing.

gmwersTaskIds :: Lens' GetMaintenanceWindowExecutionResponse [Text]

The ID of the task executions from the Maintenance Window execution.

gmwersResponseStatus :: Lens' GetMaintenanceWindowExecutionResponse Int

- | The response status code.