Network.AWS.SSM.GetMaintenanceWindowExecutionTask

Description

Retrieves the details about a specific task executed as part of a Maintenance Window execution.

Creating a Request

getMaintenanceWindowExecutionTask

Arguments

:: Text
gmwetWindowExecutionId

-> Text
gmwetTaskId

-> GetMaintenanceWindowExecutionTask

Creates a value of GetMaintenanceWindowExecutionTask with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gmwetWindowExecutionId - The ID of the Maintenance Window execution that includes the task.

gmwetTaskId - The ID of the specific task execution in the Maintenance Window task that should be retrieved.

data GetMaintenanceWindowExecutionTask

See: getMaintenanceWindowExecutionTask smart constructor.

Request Lenses

gmwetWindowExecutionId :: Lens' GetMaintenanceWindowExecutionTask Text

The ID of the Maintenance Window execution that includes the task.

gmwetTaskId :: Lens' GetMaintenanceWindowExecutionTask Text

The ID of the specific task execution in the Maintenance Window task that should be retrieved.

Destructuring the Response

getMaintenanceWindowExecutionTaskResponse

Arguments

:: Int
gmwetrsResponseStatus

-> GetMaintenanceWindowExecutionTaskResponse

Creates a value of GetMaintenanceWindowExecutionTaskResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gmwetrsStatus - The status of the task.

gmwetrsTaskParameters - The parameters passed to the task when it was executed. The map has the following format: Key: string, between 1 and 255 characters Value: an array of strings, each string is between 1 and 255 characters

gmwetrsTaskExecutionId - The ID of the specific task execution in the Maintenance Window task that was retrieved.

gmwetrsPriority - The priority of the task.

gmwetrsStartTime - The time the task execution started.

gmwetrsTaskARN - The ARN of the executed task.

gmwetrsWindowExecutionId - The ID of the Maintenance Window execution that includes the task.

gmwetrsStatusDetails - The details explaining the Status. Only available for certain status values.

gmwetrsMaxErrors - The defined maximum number of task execution errors allowed before scheduling of the task execution would have been stopped.

gmwetrsEndTime - The time the task execution completed.

gmwetrsType - The type of task executed.

gmwetrsMaxConcurrency - The defined maximum number of task executions that could be run in parallel.

gmwetrsServiceRole - The role that was assumed when executing the task.

gmwetrsResponseStatus - -- | The response status code.

data GetMaintenanceWindowExecutionTaskResponse

See: getMaintenanceWindowExecutionTaskResponse smart constructor.

Response Lenses

gmwetrsStatus :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe MaintenanceWindowExecutionStatus)

The status of the task.

gmwetrsTaskParameters :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe [HashMap Text MaintenanceWindowTaskParameterValueExpression])

The parameters passed to the task when it was executed. The map has the following format: Key: string, between 1 and 255 characters Value: an array of strings, each string is between 1 and 255 characters

gmwetrsTaskExecutionId :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe Text)

The ID of the specific task execution in the Maintenance Window task that was retrieved.

gmwetrsPriority :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe Natural)

The priority of the task.

gmwetrsStartTime :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe UTCTime)

The time the task execution started.

gmwetrsTaskARN :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe Text)

The ARN of the executed task.

gmwetrsWindowExecutionId :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe Text)

The ID of the Maintenance Window execution that includes the task.

gmwetrsStatusDetails :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe Text)

The details explaining the Status. Only available for certain status values.

gmwetrsMaxErrors :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe Text)

The defined maximum number of task execution errors allowed before scheduling of the task execution would have been stopped.

gmwetrsEndTime :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe UTCTime)

The time the task execution completed.

gmwetrsType :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe MaintenanceWindowTaskType)

The type of task executed.

gmwetrsMaxConcurrency :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe Text)

The defined maximum number of task executions that could be run in parallel.

gmwetrsServiceRole :: Lens' GetMaintenanceWindowExecutionTaskResponse (Maybe Text)

The role that was assumed when executing the task.

gmwetrsResponseStatus :: Lens' GetMaintenanceWindowExecutionTaskResponse Int

- | The response status code.