Network.AWS.SSM.GetMaintenanceWindowExecutionTaskInvocation

Description

Retrieves a task invocation. A task invocation is a specific task executing on a specific target. Maintenance Windows report status for all invocations.

Creating a Request

getMaintenanceWindowExecutionTaskInvocation

Arguments

:: Text
gmwetiWindowExecutionId

-> Text
gmwetiTaskId

-> Text
gmwetiInvocationId

-> GetMaintenanceWindowExecutionTaskInvocation

Creates a value of GetMaintenanceWindowExecutionTaskInvocation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gmwetiWindowExecutionId - The ID of the Maintenance Window execution the task is part of.

gmwetiTaskId - The ID of the specific task in the Maintenance Window task that should be retrieved.

gmwetiInvocationId - The invocation ID to retrieve.

data GetMaintenanceWindowExecutionTaskInvocation

See: getMaintenanceWindowExecutionTaskInvocation smart constructor.

Request Lenses

gmwetiWindowExecutionId :: Lens' GetMaintenanceWindowExecutionTaskInvocation Text

The ID of the Maintenance Window execution the task is part of.

gmwetiTaskId :: Lens' GetMaintenanceWindowExecutionTaskInvocation Text

The ID of the specific task in the Maintenance Window task that should be retrieved.

gmwetiInvocationId :: Lens' GetMaintenanceWindowExecutionTaskInvocation Text

The invocation ID to retrieve.

Destructuring the Response

getMaintenanceWindowExecutionTaskInvocationResponse

Arguments

:: Int
gmwetirsResponseStatus

-> GetMaintenanceWindowExecutionTaskInvocationResponse

Creates a value of GetMaintenanceWindowExecutionTaskInvocationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gmwetirsStatus - The task status for an invocation.

gmwetirsExecutionId - The execution ID.

gmwetirsTaskExecutionId - The task execution ID.

gmwetirsStartTime - The time the task started executing on the target.

gmwetirsInvocationId - The invocation ID.

gmwetirsOwnerInformation - User-provided value that will be included in any CloudWatch events raised while running tasks for these targets in this Maintenance Window.

gmwetirsTaskType - Retrieves the task type for a Maintenance Window. Task types include the following: LAMBDA, STEP_FUNCTION, AUTOMATION, RUN_COMMAND.

gmwetirsWindowTargetId - The Maintenance Window target ID.

gmwetirsWindowExecutionId - The Maintenance Window execution ID.

gmwetirsStatusDetails - The details explaining the status. Details are only available for certain status values.

gmwetirsEndTime - The time the task finished executing on the target.

gmwetirsParameters - The parameters used at the time the task executed.

gmwetirsResponseStatus - -- | The response status code.

data GetMaintenanceWindowExecutionTaskInvocationResponse

See: getMaintenanceWindowExecutionTaskInvocationResponse smart constructor.

Response Lenses

gmwetirsStatus :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe MaintenanceWindowExecutionStatus)

The task status for an invocation.

gmwetirsExecutionId :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe Text)

The execution ID.

gmwetirsTaskExecutionId :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe Text)

The task execution ID.

gmwetirsStartTime :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe UTCTime)

The time the task started executing on the target.

gmwetirsInvocationId :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe Text)

The invocation ID.

gmwetirsOwnerInformation :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe Text)

User-provided value that will be included in any CloudWatch events raised while running tasks for these targets in this Maintenance Window.

gmwetirsTaskType :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe MaintenanceWindowTaskType)

Retrieves the task type for a Maintenance Window. Task types include the following: LAMBDA, STEP_FUNCTION, AUTOMATION, RUN_COMMAND.

gmwetirsWindowTargetId :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe Text)

The Maintenance Window target ID.

gmwetirsWindowExecutionId :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe Text)

The Maintenance Window execution ID.

gmwetirsStatusDetails :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe Text)

The details explaining the status. Details are only available for certain status values.

gmwetirsEndTime :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe UTCTime)

The time the task finished executing on the target.

gmwetirsParameters :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe Text)

The parameters used at the time the task executed.

gmwetirsResponseStatus :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse Int

- | The response status code.