Network.AWS.SSM.GetMaintenanceWindowTask

Description

Lists the tasks in a Maintenance Window.

Creating a Request

getMaintenanceWindowTask

Arguments

:: Text
gmwtWindowId

-> Text
gmwtWindowTaskId

-> GetMaintenanceWindowTask

Creates a value of GetMaintenanceWindowTask with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gmwtWindowId - The Maintenance Window ID that includes the task to retrieve.

gmwtWindowTaskId - The Maintenance Window task ID to retrieve.

data GetMaintenanceWindowTask

See: getMaintenanceWindowTask smart constructor.

Request Lenses

gmwtWindowId :: Lens' GetMaintenanceWindowTask Text

The Maintenance Window ID that includes the task to retrieve.

gmwtWindowTaskId :: Lens' GetMaintenanceWindowTask Text

The Maintenance Window task ID to retrieve.

Destructuring the Response

getMaintenanceWindowTaskResponse

Arguments

:: Int
gmwtrsResponseStatus

-> GetMaintenanceWindowTaskResponse

Creates a value of GetMaintenanceWindowTaskResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gmwtrsServiceRoleARN - The IAM service role to assume during task execution.

gmwtrsWindowTaskId - The retrieved Maintenance Window task ID.

gmwtrsTaskParameters - The parameters to pass to the task when it executes.

gmwtrsPriority - The priority of the task when it executes. The lower the number, the higher the priority. Tasks that have the same priority are scheduled in parallel.

gmwtrsTaskType - The type of task to execute.

gmwtrsTaskARN - TaskArn is the resource that the task used during execution. For RUN_COMMAND and AUTOMATION task types, the TaskArn is the SSM Document NameARN. For LAMBDA tasks, TaskArn is the Function NameARN. For STEP_FUNCTION tasks, the TaskArn is the State Machine ARN.

gmwtrsMaxErrors - The maximum number of errors allowed before the task stops being scheduled.

gmwtrsTaskInvocationParameters - The parameters to pass to the task when it executes.

gmwtrsName - The retrieved task name.

gmwtrsTargets - The targets where the task should execute.

gmwtrsLoggingInfo - The location in Amazon S3 where the task results will be logged.

gmwtrsDescription - The retrieved task description.

gmwtrsMaxConcurrency - The maximum number of targets allowed to run this task in parallel.

gmwtrsWindowId - The retrieved Maintenance Window ID.

gmwtrsResponseStatus - -- | The response status code.

data GetMaintenanceWindowTaskResponse

See: getMaintenanceWindowTaskResponse smart constructor.

Response Lenses

gmwtrsServiceRoleARN :: Lens' GetMaintenanceWindowTaskResponse (Maybe Text)

The IAM service role to assume during task execution.

gmwtrsWindowTaskId :: Lens' GetMaintenanceWindowTaskResponse (Maybe Text)

The retrieved Maintenance Window task ID.

gmwtrsTaskParameters :: Lens' GetMaintenanceWindowTaskResponse (Maybe (HashMap Text MaintenanceWindowTaskParameterValueExpression))

The parameters to pass to the task when it executes.

gmwtrsPriority :: Lens' GetMaintenanceWindowTaskResponse (Maybe Natural)

The priority of the task when it executes. The lower the number, the higher the priority. Tasks that have the same priority are scheduled in parallel.

gmwtrsTaskType :: Lens' GetMaintenanceWindowTaskResponse (Maybe MaintenanceWindowTaskType)

The type of task to execute.

gmwtrsTaskARN :: Lens' GetMaintenanceWindowTaskResponse (Maybe Text)

TaskArn is the resource that the task used during execution. For RUN_COMMAND and AUTOMATION task types, the TaskArn is the SSM Document NameARN. For LAMBDA tasks, TaskArn is the Function NameARN. For STEP_FUNCTION tasks, the TaskArn is the State Machine ARN.

gmwtrsMaxErrors :: Lens' GetMaintenanceWindowTaskResponse (Maybe Text)

The maximum number of errors allowed before the task stops being scheduled.

gmwtrsTaskInvocationParameters :: Lens' GetMaintenanceWindowTaskResponse (Maybe MaintenanceWindowTaskInvocationParameters)

The parameters to pass to the task when it executes.

gmwtrsName :: Lens' GetMaintenanceWindowTaskResponse (Maybe Text)

The retrieved task name.

gmwtrsTargets :: Lens' GetMaintenanceWindowTaskResponse [Target]

The targets where the task should execute.

gmwtrsLoggingInfo :: Lens' GetMaintenanceWindowTaskResponse (Maybe LoggingInfo)

The location in Amazon S3 where the task results will be logged.

gmwtrsDescription :: Lens' GetMaintenanceWindowTaskResponse (Maybe Text)

The retrieved task description.

gmwtrsMaxConcurrency :: Lens' GetMaintenanceWindowTaskResponse (Maybe Text)

The maximum number of targets allowed to run this task in parallel.

gmwtrsWindowId :: Lens' GetMaintenanceWindowTaskResponse (Maybe Text)

The retrieved Maintenance Window ID.

gmwtrsResponseStatus :: Lens' GetMaintenanceWindowTaskResponse Int

- | The response status code.