Network.AWS.SSM.GetParametersByPath

Description

Retrieve parameters in a specific hierarchy. For more information, see Working with Systems Manager Parameters .

Request results are returned on a best-effort basis. If you specify MaxResults in the request, the response includes information up to the limit specified. The number of items returned, however, can be between zero and the value of MaxResults . If the service reaches an internal limit while processing the results, it stops the operation and returns the matching values up to that point and a NextToken . You can specify the NextToken in a subsequent call to get the next set of results.

Creating a Request

getParametersByPath

Arguments

:: Text
gpbpPath

-> GetParametersByPath

Creates a value of GetParametersByPath with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gpbpWithDecryption - Retrieve all parameters in a hierarchy with their value decrypted.

gpbpParameterFilters - Filters to limit the request results.

gpbpNextToken - A token to start the list. Use this token to get the next set of results.

gpbpRecursive - Retrieve all parameters within a hierarchy.

gpbpMaxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

gpbpPath - The hierarchy for the parameter. Hierarchies start with a forward slash () and end with the parameter name. A hierarchy can have a maximum of five levels. Examples: EnvironmentTestDBString003 FinanceProdIADOSWinServ2016license15

data GetParametersByPath

See: getParametersByPath smart constructor.

Request Lenses

gpbpWithDecryption :: Lens' GetParametersByPath (Maybe Bool)

Retrieve all parameters in a hierarchy with their value decrypted.

gpbpParameterFilters :: Lens' GetParametersByPath [ParameterStringFilter]

Filters to limit the request results.

gpbpNextToken :: Lens' GetParametersByPath (Maybe Text)

A token to start the list. Use this token to get the next set of results.

gpbpRecursive :: Lens' GetParametersByPath (Maybe Bool)

Retrieve all parameters within a hierarchy.

gpbpMaxResults :: Lens' GetParametersByPath (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

gpbpPath :: Lens' GetParametersByPath Text

The hierarchy for the parameter. Hierarchies start with a forward slash () and end with the parameter name. A hierarchy can have a maximum of five levels. Examples: EnvironmentTestDBString003 FinanceProdIADOSWinServ2016license15

Destructuring the Response

getParametersByPathResponse

Arguments

:: Int
gpbprsResponseStatus

-> GetParametersByPathResponse

Creates a value of GetParametersByPathResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gpbprsNextToken - The token for the next set of items to return. Use this token to get the next set of results.

gpbprsParameters - A list of parameters found in the specified hierarchy.

gpbprsResponseStatus - -- | The response status code.

data GetParametersByPathResponse

See: getParametersByPathResponse smart constructor.

Response Lenses

gpbprsNextToken :: Lens' GetParametersByPathResponse (Maybe Text)

The token for the next set of items to return. Use this token to get the next set of results.

gpbprsParameters :: Lens' GetParametersByPathResponse [Parameter]

A list of parameters found in the specified hierarchy.

gpbprsResponseStatus :: Lens' GetParametersByPathResponse Int

- | The response status code.