Network.AWS.SSM.GetPatchBaseline

Description

Retrieves information about a patch baseline.

Creating a Request

getPatchBaseline

Arguments

:: Text
gpbBaselineId

-> GetPatchBaseline

Creates a value of GetPatchBaseline with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gpbBaselineId - The ID of the patch baseline to retrieve.

data GetPatchBaseline

See: getPatchBaseline smart constructor.

Request Lenses

gpbBaselineId :: Lens' GetPatchBaseline Text

The ID of the patch baseline to retrieve.

Destructuring the Response

getPatchBaselineResponse

Arguments

:: Int
gpbrsResponseStatus

-> GetPatchBaselineResponse

Creates a value of GetPatchBaselineResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gpbrsApprovalRules - A set of rules used to include patches in the baseline.

gpbrsOperatingSystem - Returns the operating system specified for the patch baseline.

gpbrsGlobalFilters - A set of global filters used to exclude patches from the baseline.

gpbrsApprovedPatchesComplianceLevel - Returns the specified compliance severity level for approved patches in the patch baseline.

gpbrsApprovedPatches - A list of explicitly approved patches for the baseline.

gpbrsRejectedPatches - A list of explicitly rejected patches for the baseline.

gpbrsCreatedDate - The date the patch baseline was created.

gpbrsName - The name of the patch baseline.

gpbrsPatchGroups - Patch groups included in the patch baseline.

gpbrsModifiedDate - The date the patch baseline was last modified.

gpbrsDescription - A description of the patch baseline.

gpbrsBaselineId - The ID of the retrieved patch baseline.

gpbrsResponseStatus - -- | The response status code.

data GetPatchBaselineResponse

See: getPatchBaselineResponse smart constructor.

Response Lenses

gpbrsApprovalRules :: Lens' GetPatchBaselineResponse (Maybe PatchRuleGroup)

A set of rules used to include patches in the baseline.

gpbrsOperatingSystem :: Lens' GetPatchBaselineResponse (Maybe OperatingSystem)

Returns the operating system specified for the patch baseline.

gpbrsGlobalFilters :: Lens' GetPatchBaselineResponse (Maybe PatchFilterGroup)

A set of global filters used to exclude patches from the baseline.

gpbrsApprovedPatchesComplianceLevel :: Lens' GetPatchBaselineResponse (Maybe PatchComplianceLevel)

Returns the specified compliance severity level for approved patches in the patch baseline.

gpbrsApprovedPatches :: Lens' GetPatchBaselineResponse [Text]

A list of explicitly approved patches for the baseline.

gpbrsRejectedPatches :: Lens' GetPatchBaselineResponse [Text]

A list of explicitly rejected patches for the baseline.

gpbrsCreatedDate :: Lens' GetPatchBaselineResponse (Maybe UTCTime)

The date the patch baseline was created.

gpbrsName :: Lens' GetPatchBaselineResponse (Maybe Text)

The name of the patch baseline.

gpbrsPatchGroups :: Lens' GetPatchBaselineResponse [Text]

Patch groups included in the patch baseline.

gpbrsModifiedDate :: Lens' GetPatchBaselineResponse (Maybe UTCTime)

The date the patch baseline was last modified.

gpbrsDescription :: Lens' GetPatchBaselineResponse (Maybe Text)

A description of the patch baseline.

gpbrsBaselineId :: Lens' GetPatchBaselineResponse (Maybe Text)

The ID of the retrieved patch baseline.

gpbrsResponseStatus :: Lens' GetPatchBaselineResponse Int

- | The response status code.