Network.AWS.SSM.GetPatchBaselineForPatchGroup

Description

Retrieves the patch baseline that should be used for the specified patch group.

Creating a Request

getPatchBaselineForPatchGroup

Arguments

:: Text
gpbfpgPatchGroup

-> GetPatchBaselineForPatchGroup

Creates a value of GetPatchBaselineForPatchGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gpbfpgOperatingSystem - Returns he operating system rule specified for patch groups using the patch baseline.

gpbfpgPatchGroup - The name of the patch group whose patch baseline should be retrieved.

data GetPatchBaselineForPatchGroup

See: getPatchBaselineForPatchGroup smart constructor.

Request Lenses

gpbfpgOperatingSystem :: Lens' GetPatchBaselineForPatchGroup (Maybe OperatingSystem)

Returns he operating system rule specified for patch groups using the patch baseline.

gpbfpgPatchGroup :: Lens' GetPatchBaselineForPatchGroup Text

The name of the patch group whose patch baseline should be retrieved.

Destructuring the Response

getPatchBaselineForPatchGroupResponse

Arguments

:: Int
gpbfpgrsResponseStatus

-> GetPatchBaselineForPatchGroupResponse

Creates a value of GetPatchBaselineForPatchGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gpbfpgrsOperatingSystem - The operating system rule specified for patch groups using the patch baseline.

gpbfpgrsBaselineId - The ID of the patch baseline that should be used for the patch group.

gpbfpgrsPatchGroup - The name of the patch group.

gpbfpgrsResponseStatus - -- | The response status code.

data GetPatchBaselineForPatchGroupResponse

See: getPatchBaselineForPatchGroupResponse smart constructor.

Response Lenses

gpbfpgrsOperatingSystem :: Lens' GetPatchBaselineForPatchGroupResponse (Maybe OperatingSystem)

The operating system rule specified for patch groups using the patch baseline.

gpbfpgrsBaselineId :: Lens' GetPatchBaselineForPatchGroupResponse (Maybe Text)

The ID of the patch baseline that should be used for the patch group.

gpbfpgrsPatchGroup :: Lens' GetPatchBaselineForPatchGroupResponse (Maybe Text)

The name of the patch group.

gpbfpgrsResponseStatus :: Lens' GetPatchBaselineForPatchGroupResponse Int

- | The response status code.