Network.AWS.SSM.ListAssociations

Description

Lists the associations for the specified configuration document or instance.

http://docs.aws.amazon.com/ssm/latest/APIReference/API_ListAssociations.html

Request

data ListAssociations

Request constructor

listAssociations

Arguments

:: NonEmpty AssociationFilter
laAssociationFilterList

-> ListAssociations

ListAssociations constructor.

The fields accessible through corresponding lenses are:

laAssociationFilterList :: NonEmpty AssociationFilter

laMaxResults :: Maybe Natural

laNextToken :: Maybe Text

Request lenses

laAssociationFilterList :: Lens' ListAssociations (NonEmpty AssociationFilter)

One or more filters. Use a filter to return a more specific list of results.

laMaxResults :: Lens' ListAssociations (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

laNextToken :: Lens' ListAssociations (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

Response

data ListAssociationsResponse

Response constructor

listAssociationsResponse :: ListAssociationsResponse

ListAssociationsResponse constructor.

The fields accessible through corresponding lenses are:

larAssociations :: [Association]

larNextToken :: Maybe Text

Response lenses

larAssociations :: Lens' ListAssociationsResponse [Association]

The associations.

larNextToken :: Lens' ListAssociationsResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.