Network.AWS.SSM.ListComplianceItems

Description

For a specified resource ID, this API returns a list of compliance statuses for different resource types. Currently, you can only specify one resource ID per call. List results depend on the criteria specified in the filter.

Creating a Request

listComplianceItems :: ListComplianceItems

Creates a value of ListComplianceItems with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lResourceIds - The ID for the resources from which you want to get compliance information. Currently, you can only specify one resource ID.

lFilters - One or more compliance filters. Use a filter to return a more specific list of results.

lNextToken - A token to start the list. Use this token to get the next set of results.

lMaxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

lResourceTypes - The type of resource from which you want to get compliance information. Currently, the only supported resource type is ManagedInstance .

data ListComplianceItems

See: listComplianceItems smart constructor.

Request Lenses

lResourceIds :: Lens' ListComplianceItems (Maybe (NonEmpty Text))

The ID for the resources from which you want to get compliance information. Currently, you can only specify one resource ID.

lFilters :: Lens' ListComplianceItems [ComplianceStringFilter]

One or more compliance filters. Use a filter to return a more specific list of results.

lNextToken :: Lens' ListComplianceItems (Maybe Text)

A token to start the list. Use this token to get the next set of results.

lMaxResults :: Lens' ListComplianceItems (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

lResourceTypes :: Lens' ListComplianceItems (Maybe (NonEmpty Text))

The type of resource from which you want to get compliance information. Currently, the only supported resource type is ManagedInstance .

Destructuring the Response

listComplianceItemsResponse

Arguments

:: Int
lcirsResponseStatus

-> ListComplianceItemsResponse

Creates a value of ListComplianceItemsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcirsComplianceItems - A list of compliance information for the specified resource ID.

lcirsNextToken - The token for the next set of items to return. Use this token to get the next set of results.

lcirsResponseStatus - -- | The response status code.

data ListComplianceItemsResponse

See: listComplianceItemsResponse smart constructor.

Response Lenses

lcirsComplianceItems :: Lens' ListComplianceItemsResponse [ComplianceItem]

A list of compliance information for the specified resource ID.

lcirsNextToken :: Lens' ListComplianceItemsResponse (Maybe Text)

The token for the next set of items to return. Use this token to get the next set of results.

lcirsResponseStatus :: Lens' ListComplianceItemsResponse Int

- | The response status code.