Network.AWS.SSM.ListDocuments

Description

Describes one or more of your configuration documents.

http://docs.aws.amazon.com/ssm/latest/APIReference/API_ListDocuments.html

Request

data ListDocuments

Request constructor

listDocuments

Arguments

:: NonEmpty DocumentFilter
ldDocumentFilterList

-> ListDocuments

ListDocuments constructor.

The fields accessible through corresponding lenses are:

ldDocumentFilterList :: NonEmpty DocumentFilter

ldMaxResults :: Maybe Natural

ldNextToken :: Maybe Text

Request lenses

ldDocumentFilterList :: Lens' ListDocuments (NonEmpty DocumentFilter)

One or more filters. Use a filter to return a more specific list of results.

ldMaxResults :: Lens' ListDocuments (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

ldNextToken :: Lens' ListDocuments (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

Response

data ListDocumentsResponse

Response constructor

listDocumentsResponse :: ListDocumentsResponse

ListDocumentsResponse constructor.

The fields accessible through corresponding lenses are:

ldrDocumentIdentifiers :: [DocumentIdentifier]

ldrNextToken :: Maybe Text

Response lenses

ldrDocumentIdentifiers :: Lens' ListDocumentsResponse [DocumentIdentifier]

The names of the configuration documents.

ldrNextToken :: Lens' ListDocumentsResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.