Network.AWS.SSM.ListResourceComplianceSummaries

Description

Returns a resource-level summary count. The summary includes information about compliant and non-compliant statuses and detailed compliance-item severity counts, according to the filter criteria you specify.

Creating a Request

listResourceComplianceSummaries :: ListResourceComplianceSummaries

Creates a value of ListResourceComplianceSummaries with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrcsFilters - One or more filters. Use a filter to return a more specific list of results.

lrcsNextToken - A token to start the list. Use this token to get the next set of results.

lrcsMaxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

data ListResourceComplianceSummaries

See: listResourceComplianceSummaries smart constructor.

Request Lenses

lrcsFilters :: Lens' ListResourceComplianceSummaries [ComplianceStringFilter]

One or more filters. Use a filter to return a more specific list of results.

lrcsNextToken :: Lens' ListResourceComplianceSummaries (Maybe Text)

A token to start the list. Use this token to get the next set of results.

lrcsMaxResults :: Lens' ListResourceComplianceSummaries (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

Destructuring the Response

listResourceComplianceSummariesResponse

Arguments

:: Int
lrcsrsResponseStatus

-> ListResourceComplianceSummariesResponse

Creates a value of ListResourceComplianceSummariesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrcsrsResourceComplianceSummaryItems - A summary count for specified or targeted managed instances. Summary count includes information about compliant and non-compliant State Manager associations, patch statuses, or custom items according to the filter criteria you specify.

lrcsrsNextToken - The token for the next set of items to return. Use this token to get the next set of results.

lrcsrsResponseStatus - -- | The response status code.

data ListResourceComplianceSummariesResponse

See: listResourceComplianceSummariesResponse smart constructor.

Response Lenses

lrcsrsResourceComplianceSummaryItems :: Lens' ListResourceComplianceSummariesResponse [ResourceComplianceSummaryItem]

A summary count for specified or targeted managed instances. Summary count includes information about compliant and non-compliant State Manager associations, patch statuses, or custom items according to the filter criteria you specify.

lrcsrsNextToken :: Lens' ListResourceComplianceSummariesResponse (Maybe Text)

The token for the next set of items to return. Use this token to get the next set of results.

lrcsrsResponseStatus :: Lens' ListResourceComplianceSummariesResponse Int

- | The response status code.