Network.AWS.SSM.ModifyDocumentPermission

Description

Shares a Systems Manager document publicly or privately. If you share a document privately, you must specify the AWS user account IDs for those people who can use the document. If you share a document publicly, you must specify All as the account ID.

Creating a Request

modifyDocumentPermission

Arguments

:: Text
mdpName

-> DocumentPermissionType
mdpPermissionType

-> ModifyDocumentPermission

Creates a value of ModifyDocumentPermission with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mdpAccountIdsToAdd - The AWS user accounts that should have access to the document. The account IDs can either be a group of account IDs or All .

mdpAccountIdsToRemove - The AWS user accounts that should no longer have access to the document. The AWS user account can either be a group of account IDs or All . This action has a higher priority than AccountIdsToAdd . If you specify an account ID to add and the same ID to remove, the system removes access to the document.

mdpName - The name of the document that you want to share.

mdpPermissionType - The permission type for the document. The permission type can be Share .

data ModifyDocumentPermission

See: modifyDocumentPermission smart constructor.

Request Lenses

mdpAccountIdsToAdd :: Lens' ModifyDocumentPermission [Text]

The AWS user accounts that should have access to the document. The account IDs can either be a group of account IDs or All .

mdpAccountIdsToRemove :: Lens' ModifyDocumentPermission [Text]

The AWS user accounts that should no longer have access to the document. The AWS user account can either be a group of account IDs or All . This action has a higher priority than AccountIdsToAdd . If you specify an account ID to add and the same ID to remove, the system removes access to the document.

mdpName :: Lens' ModifyDocumentPermission Text

The name of the document that you want to share.

mdpPermissionType :: Lens' ModifyDocumentPermission DocumentPermissionType

The permission type for the document. The permission type can be Share .

Destructuring the Response

modifyDocumentPermissionResponse

Arguments

:: Int
mdprsResponseStatus

-> ModifyDocumentPermissionResponse

Creates a value of ModifyDocumentPermissionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mdprsResponseStatus - -- | The response status code.

data ModifyDocumentPermissionResponse

See: modifyDocumentPermissionResponse smart constructor.

Response Lenses

mdprsResponseStatus :: Lens' ModifyDocumentPermissionResponse Int

- | The response status code.