Network.AWS.SSM.PutInventory

Description

Bulk update custom inventory items on one more instance. The request adds an inventory item, if it doesn't already exist, or updates an inventory item, if it does exist.

Creating a Request

putInventory

Arguments

:: Text
piInstanceId

-> NonEmpty InventoryItem
piItems

-> PutInventory

Creates a value of PutInventory with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

piInstanceId - One or more instance IDs where you want to add or update inventory items.

piItems - The inventory items that you want to add or update on instances.

data PutInventory

See: putInventory smart constructor.

Request Lenses

piInstanceId :: Lens' PutInventory Text

One or more instance IDs where you want to add or update inventory items.

piItems :: Lens' PutInventory (NonEmpty InventoryItem)

The inventory items that you want to add or update on instances.

Destructuring the Response

putInventoryResponse

Arguments

:: Int
pirsResponseStatus

-> PutInventoryResponse

Creates a value of PutInventoryResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pirsResponseStatus - -- | The response status code.

data PutInventoryResponse

See: putInventoryResponse smart constructor.

Response Lenses

pirsResponseStatus :: Lens' PutInventoryResponse Int

- | The response status code.