Network.AWS.SSM.PutParameter

Description

Add one or more parameters to the system.

Creating a Request

putParameter

Arguments

:: Text
ppName

-> Text
ppValue

-> ParameterType
ppType

-> PutParameter

Creates a value of PutParameter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ppKeyId - The KMS Key ID that you want to use to encrypt a parameter when you choose the SecureString data type. If you don't specify a key ID, the system uses the default key associated with your AWS account.

ppAllowedPattern - A regular expression used to validate the parameter value. For example, for String types with values restricted to numbers, you can specify the following: AllowedPattern=^d+$

ppOverwrite - Overwrite an existing parameter. If not specified, will default to "false".

ppDescription - Information about the parameter that you want to add to the system

ppName - The name of the parameter that you want to add to the system.

ppValue - The parameter value that you want to add to the system.

ppType - The type of parameter that you want to add to the system.

data PutParameter

See: putParameter smart constructor.

Request Lenses

ppKeyId :: Lens' PutParameter (Maybe Text)

The KMS Key ID that you want to use to encrypt a parameter when you choose the SecureString data type. If you don't specify a key ID, the system uses the default key associated with your AWS account.

ppAllowedPattern :: Lens' PutParameter (Maybe Text)

A regular expression used to validate the parameter value. For example, for String types with values restricted to numbers, you can specify the following: AllowedPattern=^d+$

ppOverwrite :: Lens' PutParameter (Maybe Bool)

Overwrite an existing parameter. If not specified, will default to "false".

ppDescription :: Lens' PutParameter (Maybe Text)

Information about the parameter that you want to add to the system

ppName :: Lens' PutParameter Text

The name of the parameter that you want to add to the system.

ppValue :: Lens' PutParameter Text

The parameter value that you want to add to the system.

ppType :: Lens' PutParameter ParameterType

The type of parameter that you want to add to the system.

Destructuring the Response

putParameterResponse

Arguments

:: Int
pprsResponseStatus

-> PutParameterResponse

Creates a value of PutParameterResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pprsResponseStatus - -- | The response status code.

data PutParameterResponse

See: putParameterResponse smart constructor.

Response Lenses

pprsResponseStatus :: Lens' PutParameterResponse Int

- | The response status code.