Network.AWS.SSM.RegisterDefaultPatchBaseline

Description

Defines the default patch baseline.

Creating a Request

registerDefaultPatchBaseline

Arguments

:: Text
rdpbBaselineId

-> RegisterDefaultPatchBaseline

Creates a value of RegisterDefaultPatchBaseline with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdpbBaselineId - The ID of the patch baseline that should be the default patch baseline.

data RegisterDefaultPatchBaseline

See: registerDefaultPatchBaseline smart constructor.

Request Lenses

rdpbBaselineId :: Lens' RegisterDefaultPatchBaseline Text

The ID of the patch baseline that should be the default patch baseline.

Destructuring the Response

registerDefaultPatchBaselineResponse

Arguments

:: Int
rdpbrsResponseStatus

-> RegisterDefaultPatchBaselineResponse

Creates a value of RegisterDefaultPatchBaselineResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdpbrsBaselineId - The ID of the default patch baseline.

rdpbrsResponseStatus - -- | The response status code.

data RegisterDefaultPatchBaselineResponse

See: registerDefaultPatchBaselineResponse smart constructor.

Response Lenses

rdpbrsBaselineId :: Lens' RegisterDefaultPatchBaselineResponse (Maybe Text)

The ID of the default patch baseline.

rdpbrsResponseStatus :: Lens' RegisterDefaultPatchBaselineResponse Int

- | The response status code.