Network.AWS.SSM.RegisterPatchBaselineForPatchGroup

Description

Registers a patch baseline for a patch group.

Creating a Request

registerPatchBaselineForPatchGroup

Arguments

:: Text
rpbfpgBaselineId

-> Text
rpbfpgPatchGroup

-> RegisterPatchBaselineForPatchGroup

Creates a value of RegisterPatchBaselineForPatchGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rpbfpgBaselineId - The ID of the patch baseline to register the patch group with.

rpbfpgPatchGroup - The name of the patch group that should be registered with the patch baseline.

data RegisterPatchBaselineForPatchGroup

See: registerPatchBaselineForPatchGroup smart constructor.

Request Lenses

rpbfpgBaselineId :: Lens' RegisterPatchBaselineForPatchGroup Text

The ID of the patch baseline to register the patch group with.

rpbfpgPatchGroup :: Lens' RegisterPatchBaselineForPatchGroup Text

The name of the patch group that should be registered with the patch baseline.

Destructuring the Response

registerPatchBaselineForPatchGroupResponse

Arguments

:: Int
rpbfpgrsResponseStatus

-> RegisterPatchBaselineForPatchGroupResponse

Creates a value of RegisterPatchBaselineForPatchGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rpbfpgrsBaselineId - The ID of the patch baseline the patch group was registered with.

rpbfpgrsPatchGroup - The name of the patch group registered with the patch baseline.

rpbfpgrsResponseStatus - -- | The response status code.

data RegisterPatchBaselineForPatchGroupResponse

See: registerPatchBaselineForPatchGroupResponse smart constructor.

Response Lenses

rpbfpgrsBaselineId :: Lens' RegisterPatchBaselineForPatchGroupResponse (Maybe Text)

The ID of the patch baseline the patch group was registered with.

rpbfpgrsPatchGroup :: Lens' RegisterPatchBaselineForPatchGroupResponse (Maybe Text)

The name of the patch group registered with the patch baseline.

rpbfpgrsResponseStatus :: Lens' RegisterPatchBaselineForPatchGroupResponse Int

- | The response status code.