Network.AWS.SSM.RegisterTargetWithMaintenanceWindow

Description

Registers a target with a Maintenance Window.

Creating a Request

registerTargetWithMaintenanceWindow

Arguments

:: Text
rWindowId

-> MaintenanceWindowResourceType
rResourceType

-> RegisterTargetWithMaintenanceWindow

Creates a value of RegisterTargetWithMaintenanceWindow with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rClientToken - User-provided idempotency token.

rOwnerInformation - User-provided value that will be included in any CloudWatch events raised while running tasks for these targets in this Maintenance Window.

rName - An optional name for the target.

rDescription - An optional description for the target.

rWindowId - The ID of the Maintenance Window the target should be registered with.

rResourceType - The type of target being registered with the Maintenance Window.

rTargets - The targets (either instances or tags). Instances are specified using Key=instanceids,Values=instanceid1,instanceid2. Tags are specified using Key=name,Values=value.

data RegisterTargetWithMaintenanceWindow

See: registerTargetWithMaintenanceWindow smart constructor.

Request Lenses

rClientToken :: Lens' RegisterTargetWithMaintenanceWindow (Maybe Text)

User-provided idempotency token.

rOwnerInformation :: Lens' RegisterTargetWithMaintenanceWindow (Maybe Text)

User-provided value that will be included in any CloudWatch events raised while running tasks for these targets in this Maintenance Window.

rName :: Lens' RegisterTargetWithMaintenanceWindow (Maybe Text)

An optional name for the target.

rDescription :: Lens' RegisterTargetWithMaintenanceWindow (Maybe Text)

An optional description for the target.

rWindowId :: Lens' RegisterTargetWithMaintenanceWindow Text

The ID of the Maintenance Window the target should be registered with.

rResourceType :: Lens' RegisterTargetWithMaintenanceWindow MaintenanceWindowResourceType

The type of target being registered with the Maintenance Window.

rTargets :: Lens' RegisterTargetWithMaintenanceWindow [Target]

The targets (either instances or tags). Instances are specified using Key=instanceids,Values=instanceid1,instanceid2. Tags are specified using Key=name,Values=value.

Destructuring the Response

registerTargetWithMaintenanceWindowResponse

Arguments

:: Int
rrsResponseStatus

-> RegisterTargetWithMaintenanceWindowResponse

Creates a value of RegisterTargetWithMaintenanceWindowResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rrsWindowTargetId - The ID of the target definition in this Maintenance Window.

rrsResponseStatus - -- | The response status code.

data RegisterTargetWithMaintenanceWindowResponse

See: registerTargetWithMaintenanceWindowResponse smart constructor.

Response Lenses

rrsWindowTargetId :: Lens' RegisterTargetWithMaintenanceWindowResponse (Maybe Text)

The ID of the target definition in this Maintenance Window.

rrsResponseStatus :: Lens' RegisterTargetWithMaintenanceWindowResponse Int

- | The response status code.