Network.AWS.SSM.SendCommand

Description

Executes commands on one or more managed instances.

Creating a Request

sendCommand

Arguments

:: Text
scDocumentName

-> SendCommand

Creates a value of SendCommand with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

scServiceRoleARN - The IAM role that Systems Manager uses to send notifications.

scNotificationConfig - Configurations for sending notifications.

scDocumentHashType - Sha256 or Sha1.

scOutputS3KeyPrefix - The directory structure within the S3 bucket where the responses should be stored.

scMaxErrors - The maximum number of errors allowed without the command failing. When the command fails one more time beyond the value of MaxErrors, the systems stops sending the command to additional targets. You can specify a number like 10 or a percentage like 10%. The default value is 50. For more information about how to use MaxErrors, see Using Error Controls .

scInstanceIds - The instance IDs where the command should execute. You can specify a maximum of 50 IDs. If you prefer not to list individual instance IDs, you can instead send commands to a fleet of instances using the Targets parameter, which accepts EC2 tags. For more information about how to use Targets, see Sending Commands to a Fleet .

scOutputS3Region - (Deprecated) You can no longer specify this parameter. The system ignores it. Instead, Systems Manager automatically determines the Amazon S3 bucket region.

scTargets - (Optional) An array of search criteria that targets instances using a Key,Value combination that you specify. Targets is required if you don't provide one or more instance IDs in the call. For more information about how to use Targets, see Sending Commands to a Fleet .

scParameters - The required and optional parameters specified in the document being executed.

scDocumentHash - The Sha256 or Sha1 hash created by the system when the document was created.

scTimeoutSeconds - If this time is reached and the command has not already started executing, it will not execute.

scComment - User-specified information about the command, such as a brief description of what the command should do.

scOutputS3BucketName - The name of the S3 bucket where command execution responses should be stored.

scMaxConcurrency - (Optional) The maximum number of instances that are allowed to execute the command at the same time. You can specify a number such as 10 or a percentage such as 10%. The default value is 50. For more information about how to use MaxConcurrency, see Using Concurrency Controls .

scDocumentName - Required. The name of the Systems Manager document to execute. This can be a public document or a custom document.

data SendCommand

See: sendCommand smart constructor.

Request Lenses

scServiceRoleARN :: Lens' SendCommand (Maybe Text)

The IAM role that Systems Manager uses to send notifications.

scNotificationConfig :: Lens' SendCommand (Maybe NotificationConfig)

Configurations for sending notifications.

scDocumentHashType :: Lens' SendCommand (Maybe DocumentHashType)

Sha256 or Sha1.

scOutputS3KeyPrefix :: Lens' SendCommand (Maybe Text)

The directory structure within the S3 bucket where the responses should be stored.

scMaxErrors :: Lens' SendCommand (Maybe Text)

The maximum number of errors allowed without the command failing. When the command fails one more time beyond the value of MaxErrors, the systems stops sending the command to additional targets. You can specify a number like 10 or a percentage like 10%. The default value is 50. For more information about how to use MaxErrors, see Using Error Controls .

scInstanceIds :: Lens' SendCommand [Text]

The instance IDs where the command should execute. You can specify a maximum of 50 IDs. If you prefer not to list individual instance IDs, you can instead send commands to a fleet of instances using the Targets parameter, which accepts EC2 tags. For more information about how to use Targets, see Sending Commands to a Fleet .

scOutputS3Region :: Lens' SendCommand (Maybe Text)

(Deprecated) You can no longer specify this parameter. The system ignores it. Instead, Systems Manager automatically determines the Amazon S3 bucket region.

scTargets :: Lens' SendCommand [Target]

(Optional) An array of search criteria that targets instances using a Key,Value combination that you specify. Targets is required if you don't provide one or more instance IDs in the call. For more information about how to use Targets, see Sending Commands to a Fleet .

scParameters :: Lens' SendCommand (HashMap Text [Text])

The required and optional parameters specified in the document being executed.

scDocumentHash :: Lens' SendCommand (Maybe Text)

The Sha256 or Sha1 hash created by the system when the document was created.

scTimeoutSeconds :: Lens' SendCommand (Maybe Natural)

If this time is reached and the command has not already started executing, it will not execute.

scComment :: Lens' SendCommand (Maybe Text)

User-specified information about the command, such as a brief description of what the command should do.

scOutputS3BucketName :: Lens' SendCommand (Maybe Text)

The name of the S3 bucket where command execution responses should be stored.

scMaxConcurrency :: Lens' SendCommand (Maybe Text)

(Optional) The maximum number of instances that are allowed to execute the command at the same time. You can specify a number such as 10 or a percentage such as 10%. The default value is 50. For more information about how to use MaxConcurrency, see Using Concurrency Controls .

scDocumentName :: Lens' SendCommand Text

Required. The name of the Systems Manager document to execute. This can be a public document or a custom document.

Destructuring the Response

sendCommandResponse

Arguments

:: Int
scrsResponseStatus

-> SendCommandResponse

Creates a value of SendCommandResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

scrsCommand - The request as it was received by Systems Manager. Also provides the command ID which can be used future references to this request.

scrsResponseStatus - -- | The response status code.

data SendCommandResponse

See: sendCommandResponse smart constructor.

Response Lenses

scrsCommand :: Lens' SendCommandResponse (Maybe Command)

The request as it was received by Systems Manager. Also provides the command ID which can be used future references to this request.

scrsResponseStatus :: Lens' SendCommandResponse Int

- | The response status code.