Network.AWS.SSM.StopAutomationExecution

Description

Stop an Automation that is currently executing.

Creating a Request

stopAutomationExecution

Arguments

:: Text
saeAutomationExecutionId

-> StopAutomationExecution

Creates a value of StopAutomationExecution with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

saeAutomationExecutionId - The execution ID of the Automation to stop.

data StopAutomationExecution

See: stopAutomationExecution smart constructor.

Request Lenses

saeAutomationExecutionId :: Lens' StopAutomationExecution Text

The execution ID of the Automation to stop.

Destructuring the Response

stopAutomationExecutionResponse

Arguments

:: Int
saersResponseStatus

-> StopAutomationExecutionResponse

Creates a value of StopAutomationExecutionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

saersResponseStatus - -- | The response status code.

data StopAutomationExecutionResponse

See: stopAutomationExecutionResponse smart constructor.

Response Lenses

saersResponseStatus :: Lens' StopAutomationExecutionResponse Int

- | The response status code.