Network.AWS.SSM.UpdateAssociation

Description

Updates an association. You can only update the document version, schedule, parameters, and Amazon S3 output of an association.

Creating a Request

updateAssociation

Arguments

:: Text
uaAssociationId

-> UpdateAssociation

Creates a value of UpdateAssociation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uaScheduleExpression - The cron expression used to schedule the association that you want to update.

uaName - The name of the association document.

uaOutputLocation - An Amazon S3 bucket where you want to store the results of this request.

uaTargets - The targets of the association.

uaParameters - The parameters you want to update for the association. If you create a parameter using Parameter Store, you can reference the parameter using {{ssm:parameter-name}}

uaDocumentVersion - The document version you want update for the association.

uaAssociationId - The ID of the association you want to update.

data UpdateAssociation

See: updateAssociation smart constructor.

Request Lenses

uaScheduleExpression :: Lens' UpdateAssociation (Maybe Text)

The cron expression used to schedule the association that you want to update.

uaName :: Lens' UpdateAssociation (Maybe Text)

The name of the association document.

uaOutputLocation :: Lens' UpdateAssociation (Maybe InstanceAssociationOutputLocation)

An Amazon S3 bucket where you want to store the results of this request.

uaTargets :: Lens' UpdateAssociation [Target]

The targets of the association.

uaParameters :: Lens' UpdateAssociation (HashMap Text [Text])

The parameters you want to update for the association. If you create a parameter using Parameter Store, you can reference the parameter using {{ssm:parameter-name}}

uaDocumentVersion :: Lens' UpdateAssociation (Maybe Text)

The document version you want update for the association.

uaAssociationId :: Lens' UpdateAssociation Text

The ID of the association you want to update.

Destructuring the Response

updateAssociationResponse

Arguments

:: Int
uarsResponseStatus

-> UpdateAssociationResponse

Creates a value of UpdateAssociationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uarsAssociationDescription - The description of the association that was updated.

uarsResponseStatus - -- | The response status code.

data UpdateAssociationResponse

See: updateAssociationResponse smart constructor.

Response Lenses

uarsAssociationDescription :: Lens' UpdateAssociationResponse (Maybe AssociationDescription)

The description of the association that was updated.

uarsResponseStatus :: Lens' UpdateAssociationResponse Int

- | The response status code.