Network.AWS.SSM.UpdateMaintenanceWindow

Description

Updates an existing Maintenance Window. Only specified parameters are modified.

Creating a Request

updateMaintenanceWindow

Arguments

:: Text
umwWindowId

-> UpdateMaintenanceWindow

Creates a value of UpdateMaintenanceWindow with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

umwReplace - If you specify True, then all fields that are required by the CreateMaintenanceWindow API are also required for this API request. Optional fields that are not specified will be set to null.

umwEnabled - Whether the Maintenance Window is enabled.

umwSchedule - The schedule of the Maintenance Window in the form of a cron or rate expression.

umwName - The name of the Maintenance Window.

umwCutoff - The number of hours before the end of the Maintenance Window that Systems Manager stops scheduling new tasks for execution.

umwAllowUnassociatedTargets - Whether targets must be registered with the Maintenance Window before tasks can be defined for those targets.

umwDescription - An optional description for the update request.

umwDuration - The duration of the Maintenance Window in hours.

umwWindowId - The ID of the Maintenance Window to update.

data UpdateMaintenanceWindow

See: updateMaintenanceWindow smart constructor.

Request Lenses

umwReplace :: Lens' UpdateMaintenanceWindow (Maybe Bool)

If you specify True, then all fields that are required by the CreateMaintenanceWindow API are also required for this API request. Optional fields that are not specified will be set to null.

umwEnabled :: Lens' UpdateMaintenanceWindow (Maybe Bool)

Whether the Maintenance Window is enabled.

umwSchedule :: Lens' UpdateMaintenanceWindow (Maybe Text)

The schedule of the Maintenance Window in the form of a cron or rate expression.

umwName :: Lens' UpdateMaintenanceWindow (Maybe Text)

The name of the Maintenance Window.

umwCutoff :: Lens' UpdateMaintenanceWindow (Maybe Natural)

The number of hours before the end of the Maintenance Window that Systems Manager stops scheduling new tasks for execution.

umwAllowUnassociatedTargets :: Lens' UpdateMaintenanceWindow (Maybe Bool)

Whether targets must be registered with the Maintenance Window before tasks can be defined for those targets.

umwDescription :: Lens' UpdateMaintenanceWindow (Maybe Text)

An optional description for the update request.

umwDuration :: Lens' UpdateMaintenanceWindow (Maybe Natural)

The duration of the Maintenance Window in hours.

umwWindowId :: Lens' UpdateMaintenanceWindow Text

The ID of the Maintenance Window to update.

Destructuring the Response

updateMaintenanceWindowResponse

Arguments

:: Int
umwrsResponseStatus

-> UpdateMaintenanceWindowResponse

Creates a value of UpdateMaintenanceWindowResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

umwrsEnabled - Whether the Maintenance Window is enabled.

umwrsSchedule - The schedule of the Maintenance Window in the form of a cron or rate expression.

umwrsName - The name of the Maintenance Window.

umwrsCutoff - The number of hours before the end of the Maintenance Window that Systems Manager stops scheduling new tasks for execution.

umwrsAllowUnassociatedTargets - Whether targets must be registered with the Maintenance Window before tasks can be defined for those targets.

umwrsDescription - An optional description of the update.

umwrsDuration - The duration of the Maintenance Window in hours.

umwrsWindowId - The ID of the created Maintenance Window.

umwrsResponseStatus - -- | The response status code.

data UpdateMaintenanceWindowResponse

See: updateMaintenanceWindowResponse smart constructor.

Response Lenses

umwrsEnabled :: Lens' UpdateMaintenanceWindowResponse (Maybe Bool)

Whether the Maintenance Window is enabled.

umwrsSchedule :: Lens' UpdateMaintenanceWindowResponse (Maybe Text)

The schedule of the Maintenance Window in the form of a cron or rate expression.

umwrsName :: Lens' UpdateMaintenanceWindowResponse (Maybe Text)

The name of the Maintenance Window.

umwrsCutoff :: Lens' UpdateMaintenanceWindowResponse (Maybe Natural)

The number of hours before the end of the Maintenance Window that Systems Manager stops scheduling new tasks for execution.

umwrsAllowUnassociatedTargets :: Lens' UpdateMaintenanceWindowResponse (Maybe Bool)

Whether targets must be registered with the Maintenance Window before tasks can be defined for those targets.

umwrsDescription :: Lens' UpdateMaintenanceWindowResponse (Maybe Text)

An optional description of the update.

umwrsDuration :: Lens' UpdateMaintenanceWindowResponse (Maybe Natural)

The duration of the Maintenance Window in hours.

umwrsWindowId :: Lens' UpdateMaintenanceWindowResponse (Maybe Text)

The ID of the created Maintenance Window.

umwrsResponseStatus :: Lens' UpdateMaintenanceWindowResponse Int

- | The response status code.