Network.AWS.SSM.UpdateMaintenanceWindowTarget

Description

Modifies the target of an existing Maintenance Window. You can't change the target type, but you can change the following:

The target from being an ID target to a Tag target, or a Tag target to an ID target.

The IDs of an ID target.

The tags of a Tag target.

The Owner.

The Name.

The Description.

Also note that if a parameter is null, then the corresponding field is not modified.

Creating a Request

updateMaintenanceWindowTarget

Arguments

:: Text
uWindowId

-> Text
uWindowTargetId

-> UpdateMaintenanceWindowTarget

Creates a value of UpdateMaintenanceWindowTarget with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uReplace - If you specify True, then all fields that are required by the RegisterTargetWithMaintenanceWindow API are also required for this API request. Optional fields that are not specified will be set to null.

uOwnerInformation - User-provided value that will be included in any CloudWatch events raised while running tasks for these targets in this Maintenance Window.

uName - A name for the update.

uTargets - The targets that you want to add or replace.

uDescription - An optional description for the update.

uWindowId - The Maintenance Window ID for which you want to modify the target.

uWindowTargetId - The target ID that you want to modify.

data UpdateMaintenanceWindowTarget

See: updateMaintenanceWindowTarget smart constructor.

Request Lenses

uReplace :: Lens' UpdateMaintenanceWindowTarget (Maybe Bool)

If you specify True, then all fields that are required by the RegisterTargetWithMaintenanceWindow API are also required for this API request. Optional fields that are not specified will be set to null.

uOwnerInformation :: Lens' UpdateMaintenanceWindowTarget (Maybe Text)

User-provided value that will be included in any CloudWatch events raised while running tasks for these targets in this Maintenance Window.

uName :: Lens' UpdateMaintenanceWindowTarget (Maybe Text)

A name for the update.

uTargets :: Lens' UpdateMaintenanceWindowTarget [Target]

The targets that you want to add or replace.

uDescription :: Lens' UpdateMaintenanceWindowTarget (Maybe Text)

An optional description for the update.

uWindowId :: Lens' UpdateMaintenanceWindowTarget Text

The Maintenance Window ID for which you want to modify the target.

uWindowTargetId :: Lens' UpdateMaintenanceWindowTarget Text

The target ID that you want to modify.

Destructuring the Response

updateMaintenanceWindowTargetResponse

Arguments

:: Int
ursResponseStatus

-> UpdateMaintenanceWindowTargetResponse

Creates a value of UpdateMaintenanceWindowTargetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ursOwnerInformation - The updated owner.

ursWindowTargetId - The target ID specified in the update request.

ursName - The updated name.

ursTargets - The updated targets.

ursDescription - The updated description.

ursWindowId - The Maintenance Window ID specified in the update request.

ursResponseStatus - -- | The response status code.

data UpdateMaintenanceWindowTargetResponse

See: updateMaintenanceWindowTargetResponse smart constructor.

Response Lenses

ursOwnerInformation :: Lens' UpdateMaintenanceWindowTargetResponse (Maybe Text)

The updated owner.

ursWindowTargetId :: Lens' UpdateMaintenanceWindowTargetResponse (Maybe Text)

The target ID specified in the update request.

ursName :: Lens' UpdateMaintenanceWindowTargetResponse (Maybe Text)

The updated name.

ursTargets :: Lens' UpdateMaintenanceWindowTargetResponse [Target]

The updated targets.

ursDescription :: Lens' UpdateMaintenanceWindowTargetResponse (Maybe Text)

The updated description.

ursWindowId :: Lens' UpdateMaintenanceWindowTargetResponse (Maybe Text)

The Maintenance Window ID specified in the update request.

ursResponseStatus :: Lens' UpdateMaintenanceWindowTargetResponse Int

- | The response status code.