Network.AWS.SSM.UpdateMaintenanceWindowTask

Description

Modifies a task assigned to a Maintenance Window. You can't change the task type, but you can change the following:

The Task Arn. For example, you can change a RUN_COMMAND task from AWS-RunPowerShellScript to AWS-RunShellScript.

The service role ARN.

The task parameters.

The task priority.

The task MaxConcurrency and MaxErrors.

The log location.

If a parameter is null, then the corresponding field is not modified. Also, if you set Replace to true, then all fields required by the RegisterTaskWithMaintenanceWindow operation are required for this request. Optional fields that aren't specified are be set to null.

Creating a Request

updateMaintenanceWindowTask

Arguments

:: Text
umwtWindowId

-> Text
umwtWindowTaskId

-> UpdateMaintenanceWindowTask

Creates a value of UpdateMaintenanceWindowTask with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

umwtServiceRoleARN - The IAM service role ARN that you want to modify. The system assumes this role during task exectuion.

umwtReplace - If you specify True, then all fields that are required by the RegisterTaskWithMaintenanceWndow API are also required for this API request. Optional fields that are not specified will be set to null.

umwtTaskParameters - The parameters that you want to modify. The map has the following format: Key: string, between 1 and 255 characters Value: an array of strings, each string is between 1 and 255 characters

umwtPriority - The new task priority that you want to specify. The lower the number, the higher the priority. Tasks that have the same priority are scheduled in parallel.

umwtTaskARN - The task ARN that you want to modify.

umwtMaxErrors - The new MaxErrors value you want to specify. MaxErrors is the maximum number of errors that are allowed before the task stops being scheduled.

umwtTaskInvocationParameters - Parameters the task should use during execution. Populate only the fields that match the task type. All other fields should be empty.

umwtName - The new task name that you want to specify.

umwtTargets - The targets (either instances or tags) that you want to modify. Instances are specified using Key=instanceids,Values=instanceID_1,instanceID_2. Tags are specified using Key=tag_name,Values=tag_value.

umwtLoggingInfo - The new logging location in Amazon S3 that you want to specify.

umwtDescription - The new task description that you want to specify.

umwtMaxConcurrency - The new MaxConcurrency value you want to specify. MaxConcurrency is the number of targets that are allowed to run this task in parallel.

umwtWindowId - The Maintenance Window ID that contains the task that you want to modify.

umwtWindowTaskId - The task ID that you want to modify.

data UpdateMaintenanceWindowTask

See: updateMaintenanceWindowTask smart constructor.

Request Lenses

umwtServiceRoleARN :: Lens' UpdateMaintenanceWindowTask (Maybe Text)

The IAM service role ARN that you want to modify. The system assumes this role during task exectuion.

umwtReplace :: Lens' UpdateMaintenanceWindowTask (Maybe Bool)

If you specify True, then all fields that are required by the RegisterTaskWithMaintenanceWndow API are also required for this API request. Optional fields that are not specified will be set to null.

umwtTaskParameters :: Lens' UpdateMaintenanceWindowTask (Maybe (HashMap Text MaintenanceWindowTaskParameterValueExpression))

The parameters that you want to modify. The map has the following format: Key: string, between 1 and 255 characters Value: an array of strings, each string is between 1 and 255 characters

umwtPriority :: Lens' UpdateMaintenanceWindowTask (Maybe Natural)

The new task priority that you want to specify. The lower the number, the higher the priority. Tasks that have the same priority are scheduled in parallel.

umwtTaskARN :: Lens' UpdateMaintenanceWindowTask (Maybe Text)

The task ARN that you want to modify.

umwtMaxErrors :: Lens' UpdateMaintenanceWindowTask (Maybe Text)

The new MaxErrors value you want to specify. MaxErrors is the maximum number of errors that are allowed before the task stops being scheduled.

umwtTaskInvocationParameters :: Lens' UpdateMaintenanceWindowTask (Maybe MaintenanceWindowTaskInvocationParameters)

Parameters the task should use during execution. Populate only the fields that match the task type. All other fields should be empty.

umwtName :: Lens' UpdateMaintenanceWindowTask (Maybe Text)

The new task name that you want to specify.

umwtTargets :: Lens' UpdateMaintenanceWindowTask [Target]

The targets (either instances or tags) that you want to modify. Instances are specified using Key=instanceids,Values=instanceID_1,instanceID_2. Tags are specified using Key=tag_name,Values=tag_value.

umwtLoggingInfo :: Lens' UpdateMaintenanceWindowTask (Maybe LoggingInfo)

The new logging location in Amazon S3 that you want to specify.

umwtDescription :: Lens' UpdateMaintenanceWindowTask (Maybe Text)

The new task description that you want to specify.

umwtMaxConcurrency :: Lens' UpdateMaintenanceWindowTask (Maybe Text)

The new MaxConcurrency value you want to specify. MaxConcurrency is the number of targets that are allowed to run this task in parallel.

umwtWindowId :: Lens' UpdateMaintenanceWindowTask Text

The Maintenance Window ID that contains the task that you want to modify.

umwtWindowTaskId :: Lens' UpdateMaintenanceWindowTask Text

The task ID that you want to modify.

Destructuring the Response

updateMaintenanceWindowTaskResponse

Arguments

:: Int
umwtrsResponseStatus

-> UpdateMaintenanceWindowTaskResponse

Creates a value of UpdateMaintenanceWindowTaskResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

umwtrsServiceRoleARN - The updated service role ARN value.

umwtrsWindowTaskId - The Maintenance Window task ID that was updated.

umwtrsTaskParameters - The updated parameter values.

umwtrsPriority - The updated priority value.

umwtrsTaskARN - The updated task ARN value.

umwtrsMaxErrors - The updated MaxErrors value.

umwtrsTaskInvocationParameters - The updated parameter values.

umwtrsName - The updated task name.

umwtrsTargets - The updated target values.

umwtrsLoggingInfo - The updated logging information in Amazon S3.

umwtrsDescription - The updated task description.

umwtrsMaxConcurrency - The updated MaxConcurrency value.

umwtrsWindowId - The Maintenance Window ID that was updated.

umwtrsResponseStatus - -- | The response status code.

data UpdateMaintenanceWindowTaskResponse

See: updateMaintenanceWindowTaskResponse smart constructor.

Response Lenses

umwtrsServiceRoleARN :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe Text)

The updated service role ARN value.

umwtrsWindowTaskId :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe Text)

The Maintenance Window task ID that was updated.

umwtrsTaskParameters :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe (HashMap Text MaintenanceWindowTaskParameterValueExpression))

The updated parameter values.

umwtrsPriority :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe Natural)

The updated priority value.

umwtrsTaskARN :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe Text)

The updated task ARN value.

umwtrsMaxErrors :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe Text)

The updated MaxErrors value.

umwtrsTaskInvocationParameters :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe MaintenanceWindowTaskInvocationParameters)

The updated parameter values.

umwtrsName :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe Text)

The updated task name.

umwtrsTargets :: Lens' UpdateMaintenanceWindowTaskResponse [Target]

The updated target values.

umwtrsLoggingInfo :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe LoggingInfo)

The updated logging information in Amazon S3.

umwtrsDescription :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe Text)

The updated task description.

umwtrsMaxConcurrency :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe Text)

The updated MaxConcurrency value.

umwtrsWindowId :: Lens' UpdateMaintenanceWindowTaskResponse (Maybe Text)

The Maintenance Window ID that was updated.

umwtrsResponseStatus :: Lens' UpdateMaintenanceWindowTaskResponse Int

- | The response status code.