Network.AWS.SSM.UpdateManagedInstanceRole

Description

Assigns or changes an Amazon Identity and Access Management (IAM) role to the managed instance.

Creating a Request

updateManagedInstanceRole

Arguments

:: Text
umirInstanceId

-> Text
umirIAMRole

-> UpdateManagedInstanceRole

Creates a value of UpdateManagedInstanceRole with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

umirInstanceId - The ID of the managed instance where you want to update the role.

umirIAMRole - The IAM role you want to assign or change.

data UpdateManagedInstanceRole

See: updateManagedInstanceRole smart constructor.

Request Lenses

umirInstanceId :: Lens' UpdateManagedInstanceRole Text

The ID of the managed instance where you want to update the role.

umirIAMRole :: Lens' UpdateManagedInstanceRole Text

The IAM role you want to assign or change.

Destructuring the Response

updateManagedInstanceRoleResponse

Arguments

:: Int
umirrsResponseStatus

-> UpdateManagedInstanceRoleResponse

Creates a value of UpdateManagedInstanceRoleResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

umirrsResponseStatus - -- | The response status code.

data UpdateManagedInstanceRoleResponse

See: updateManagedInstanceRoleResponse smart constructor.

Response Lenses

umirrsResponseStatus :: Lens' UpdateManagedInstanceRoleResponse Int

- | The response status code.