Network.AWS.SSM.UpdatePatchBaseline

Description

Modifies an existing patch baseline. Fields not specified in the request are left unchanged.

Creating a Request

updatePatchBaseline

Arguments

:: Text
upbBaselineId

-> UpdatePatchBaseline

Creates a value of UpdatePatchBaseline with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

upbApprovalRules - A set of rules used to include patches in the baseline.

upbGlobalFilters - A set of global filters used to exclude patches from the baseline.

upbApprovedPatchesComplianceLevel - Assigns a new compliance severity level to an existing patch baseline.

upbApprovedPatches - A list of explicitly approved patches for the baseline.

upbRejectedPatches - A list of explicitly rejected patches for the baseline.

upbName - The name of the patch baseline.

upbDescription - A description of the patch baseline.

upbBaselineId - The ID of the patch baseline to update.

data UpdatePatchBaseline

See: updatePatchBaseline smart constructor.

Request Lenses

upbApprovalRules :: Lens' UpdatePatchBaseline (Maybe PatchRuleGroup)

A set of rules used to include patches in the baseline.

upbGlobalFilters :: Lens' UpdatePatchBaseline (Maybe PatchFilterGroup)

A set of global filters used to exclude patches from the baseline.

upbApprovedPatchesComplianceLevel :: Lens' UpdatePatchBaseline (Maybe PatchComplianceLevel)

Assigns a new compliance severity level to an existing patch baseline.

upbApprovedPatches :: Lens' UpdatePatchBaseline [Text]

A list of explicitly approved patches for the baseline.

upbRejectedPatches :: Lens' UpdatePatchBaseline [Text]

A list of explicitly rejected patches for the baseline.

upbName :: Lens' UpdatePatchBaseline (Maybe Text)

The name of the patch baseline.

upbDescription :: Lens' UpdatePatchBaseline (Maybe Text)

A description of the patch baseline.

upbBaselineId :: Lens' UpdatePatchBaseline Text

The ID of the patch baseline to update.

Destructuring the Response

updatePatchBaselineResponse

Arguments

:: Int
upbrsResponseStatus

-> UpdatePatchBaselineResponse

Creates a value of UpdatePatchBaselineResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

upbrsApprovalRules - A set of rules used to include patches in the baseline.

upbrsOperatingSystem - The operating system rule used by the updated patch baseline.

upbrsGlobalFilters - A set of global filters used to exclude patches from the baseline.

upbrsApprovedPatchesComplianceLevel - The compliance severity level assigned to the patch baseline after the update completed.

upbrsApprovedPatches - A list of explicitly approved patches for the baseline.

upbrsRejectedPatches - A list of explicitly rejected patches for the baseline.

upbrsCreatedDate - The date when the patch baseline was created.

upbrsName - The name of the patch baseline.

upbrsModifiedDate - The date when the patch baseline was last modified.

upbrsDescription - A description of the Patch Baseline.

upbrsBaselineId - The ID of the deleted patch baseline.

upbrsResponseStatus - -- | The response status code.

data UpdatePatchBaselineResponse

See: updatePatchBaselineResponse smart constructor.

Response Lenses

upbrsApprovalRules :: Lens' UpdatePatchBaselineResponse (Maybe PatchRuleGroup)

A set of rules used to include patches in the baseline.

upbrsOperatingSystem :: Lens' UpdatePatchBaselineResponse (Maybe OperatingSystem)

The operating system rule used by the updated patch baseline.

upbrsGlobalFilters :: Lens' UpdatePatchBaselineResponse (Maybe PatchFilterGroup)

A set of global filters used to exclude patches from the baseline.

upbrsApprovedPatchesComplianceLevel :: Lens' UpdatePatchBaselineResponse (Maybe PatchComplianceLevel)

The compliance severity level assigned to the patch baseline after the update completed.

upbrsApprovedPatches :: Lens' UpdatePatchBaselineResponse [Text]

A list of explicitly approved patches for the baseline.

upbrsRejectedPatches :: Lens' UpdatePatchBaselineResponse [Text]

A list of explicitly rejected patches for the baseline.

upbrsCreatedDate :: Lens' UpdatePatchBaselineResponse (Maybe UTCTime)

The date when the patch baseline was created.

upbrsName :: Lens' UpdatePatchBaselineResponse (Maybe Text)

The name of the patch baseline.

upbrsModifiedDate :: Lens' UpdatePatchBaselineResponse (Maybe UTCTime)

The date when the patch baseline was last modified.

upbrsDescription :: Lens' UpdatePatchBaselineResponse (Maybe Text)

A description of the Patch Baseline.

upbrsBaselineId :: Lens' UpdatePatchBaselineResponse (Maybe Text)

The ID of the deleted patch baseline.

upbrsResponseStatus :: Lens' UpdatePatchBaselineResponse Int

- | The response status code.