Network.AWS.SSM.CreateAssociationBatch

Description

Associates the specified Systems Manager document with the specified instances or targets.

When you associate a document with one or more instances using instance IDs or tags, the SSM Agent running on the instance processes the document and configures the instance as specified.

If you associate a document with an instance that already has an associated document, the system throws the AssociationAlreadyExists exception.

Creating a Request

createAssociationBatch

Arguments

:: NonEmpty CreateAssociationBatchRequestEntry
cabEntries

-> CreateAssociationBatch

Creates a value of CreateAssociationBatch with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cabEntries - One or more associations.

data CreateAssociationBatch

See: createAssociationBatch smart constructor.

Request Lenses

cabEntries :: Lens' CreateAssociationBatch (NonEmpty CreateAssociationBatchRequestEntry)

One or more associations.

Destructuring the Response

createAssociationBatchResponse

Arguments

:: Int
cabrsResponseStatus

-> CreateAssociationBatchResponse

Creates a value of CreateAssociationBatchResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cabrsSuccessful - Information about the associations that succeeded.

cabrsFailed - Information about the associations that failed.

cabrsResponseStatus - -- | The response status code.

data CreateAssociationBatchResponse

See: createAssociationBatchResponse smart constructor.

Response Lenses

cabrsSuccessful :: Lens' CreateAssociationBatchResponse [AssociationDescription]

Information about the associations that succeeded.

cabrsFailed :: Lens' CreateAssociationBatchResponse [FailedCreateAssociation]

Information about the associations that failed.

cabrsResponseStatus :: Lens' CreateAssociationBatchResponse Int

- | The response status code.