Network.AWS.SSM.CreateDocument

Description

Creates a Systems Manager document.

After you create a document, you can use CreateAssociation to associate it with one or more running instances.

Creating a Request

createDocument

Arguments

:: Text
cdContent

-> Text
cdName

-> CreateDocument

Creates a value of CreateDocument with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdDocumentType - The type of document to create. Valid document types include: Policy, Automation, and Command.

cdContent - A valid JSON string.

cdName - A name for the Systems Manager document.

data CreateDocument

See: createDocument smart constructor.

Request Lenses

cdDocumentType :: Lens' CreateDocument (Maybe DocumentType)

The type of document to create. Valid document types include: Policy, Automation, and Command.

cdContent :: Lens' CreateDocument Text

A valid JSON string.

cdName :: Lens' CreateDocument Text

A name for the Systems Manager document.

Destructuring the Response

createDocumentResponse

Arguments

:: Int
cdrsResponseStatus

-> CreateDocumentResponse

Creates a value of CreateDocumentResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdrsDocumentDescription - Information about the Systems Manager document.

cdrsResponseStatus - -- | The response status code.

data CreateDocumentResponse

See: createDocumentResponse smart constructor.

Response Lenses

cdrsDocumentDescription :: Lens' CreateDocumentResponse (Maybe DocumentDescription)

Information about the Systems Manager document.

cdrsResponseStatus :: Lens' CreateDocumentResponse Int

- | The response status code.