Network.AWS.SSM.ListDocuments

Description

Describes one or more of your SSM documents.

This operation returns paginated results.

Creating a Request

listDocuments :: ListDocuments

Creates a value of ListDocuments with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldDocumentFilterList - One or more filters. Use a filter to return a more specific list of results.

ldNextToken - The token for the next set of items to return. (You received this token from a previous call.)

ldMaxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

data ListDocuments

See: listDocuments smart constructor.

Request Lenses

ldDocumentFilterList :: Lens' ListDocuments (Maybe (NonEmpty DocumentFilter))

One or more filters. Use a filter to return a more specific list of results.

ldNextToken :: Lens' ListDocuments (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

ldMaxResults :: Lens' ListDocuments (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

Destructuring the Response

listDocumentsResponse

Arguments

:: Int
ldrsResponseStatus

-> ListDocumentsResponse

Creates a value of ListDocumentsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldrsDocumentIdentifiers - The names of the SSM documents.

ldrsNextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

ldrsResponseStatus - -- | The response status code.

data ListDocumentsResponse

See: listDocumentsResponse smart constructor.

Response Lenses

ldrsDocumentIdentifiers :: Lens' ListDocumentsResponse [DocumentIdentifier]

The names of the SSM documents.

ldrsNextToken :: Lens' ListDocumentsResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

ldrsResponseStatus :: Lens' ListDocumentsResponse Int

- | The response status code.